Exported functions and types

Index

ControlSystems.DelayLtiSystem

ControlSystems.DelayLtiSystem

ControlSystems.Simulator

ControlSystems.Simulator

ControlSystems.TransferFunction

Base.step

CommonSolve.solve

ControlSystems.append

ControlSystems.append

ControlSystems.balance

ControlSystems.balance_statespace

ControlSystems.balreal

ControlSystems.baltrunc

ControlSystems.bode

ControlSystems.bodeplot

ControlSystems.bodev

ControlSystems.bodev

ControlSystems.c2d

ControlSystems.c2d

ControlSystems.c2d

ControlSystems.c2d_poly2poly

ControlSystems.c2d_roots2poly

ControlSystems.c2d_x0map

ControlSystems.care

ControlSystems.covar

ControlSystems.ctrb

ControlSystems.d2c

ControlSystems.dab

ControlSystems.damp

ControlSystems.dampreport

ControlSystems.dare

ControlSystems.dcgain

ControlSystems.delaymargin

ControlSystems.diagonalize

ControlSystems.dkalman

ControlSystems.dlqr

ControlSystems.dlyap

ControlSystems.evalfr

ControlSystems.feedback

ControlSystems.feedback

ControlSystems.feedback

ControlSystems.feedback2dof

ControlSystems.feedback2dof

ControlSystems.feedback2dof

ControlSystems.freqresp

ControlSystems.freqrespv

ControlSystems.freqrespv

ControlSystems.gangoffour

ControlSystems.gangoffourplot

ControlSystems.gangofseven

ControlSystems.gram

ControlSystems.grampd

ControlSystems.hinfnorm

ControlSystems.impulse

ControlSystems.innovation_form

ControlSystems.iscontinuous

ControlSystems.isdiscrete

ControlSystems.isproper

ControlSystems.isstable

ControlSystems.kalman

ControlSystems.laglink

ControlSystems.leadlink

ControlSystems.leadlinkat

ControlSystems.leadlinkcurve

ControlSystems.lft

ControlSystems.linfnorm

ControlSystems.loopshapingPI

ControlSystems.lqr

ControlSystems.lsim

ControlSystems.lsim

ControlSystems.margin

ControlSystems.markovparam

ControlSystems.minreal

ControlSystems.minreal

ControlSystems.minreal

ControlSystems.nicholsplot

ControlSystems.nyquist

ControlSystems.nyquistplot

ControlSystems.nyquistv

ControlSystems.nyquistv

ControlSystems.observer_controller

ControlSystems.observer_predictor

ControlSystems.obsv

ControlSystems.pade

ControlSystems.pade

ControlSystems.parallel

ControlSystems.parallel

ControlSystems.pid

ControlSystems.pidplots

ControlSystems.place

ControlSystems.placePI

ControlSystems.poles

ControlSystems.prescale

ControlSystems.pzmap

ControlSystems.reduce_sys

ControlSystems.relative_gain_array

ControlSystems.relative_gain_array

ControlSystems.rgaplot

ControlSystems.rlocusplot

ControlSystems.rstc

ControlSystems.rstd

ControlSystems.series

ControlSystems.series

ControlSystems.setPlotScale

ControlSystems.sigma

ControlSystems.sigmaplot

ControlSystems.sigmav

ControlSystems.sigmav

ControlSystems.similarity_transform

ControlSystems.sminreal

ControlSystems.sminreal

ControlSystems.ss

ControlSystems.ss

ControlSystems.ssdata

ControlSystems.ssrand

ControlSystems.stabregionPID

ControlSystems.starprod

ControlSystems.tf

ControlSystems.tf

ControlSystems.tzeros

ControlSystems.zpconv

ControlSystems.zpk

ControlSystems.zpk

ControlSystems.zpkdata

LinearAlgebra.norm

Docstrings

ControlSystems.DelayLtiSystem — Type

struct DelayLtiSystem{T, S <: Real} <: LTISystem

Represents an LTISystem with internal time-delay. See ?delay for a convenience constructor.

ControlSystems.DelayLtiSystem — Method

DelayLtiSystem{T, S}(sys::StateSpace, Tau::AbstractVector{S}=Float64[]) where {T <: Number, S <: Real}

Create a delayed system by speciying both the system and time-delay vector. NOTE: if you want to create a system with simple input or output delays, use the Function delay(τ).

ControlSystems.Simulator — Type

Simulator

Fields:

P::StateSpace
f = (x,p,t) -> x
y = (x,t)   -> y

ControlSystems.Simulator — Method

Simulator(P::StateSpace, u = (x,t) -> 0)

Used to simulate continuous-time systems. See function ?solve for additional info.

Usage:

using OrdinaryDiffEq, Plots
dt             = 0.1
tfinal         = 20
t              = 0:dt:tfinal
P              = ss(tf(1,[2,1])^2)
K              = 5
reference(x,t) = [1.]
s              = Simulator(P, reference)
x0             = [0.,0]
tspan          = (0.0,tfinal)
sol            = solve(s, x0, tspan, Tsit5())
plot(t, s.y(sol, t)[:], lab="Open loop step response")

ControlSystems.TransferFunction — Method

F(s), F(omega, true), F(z, false)

Notation for frequency response evaluation.

F(s) evaluates the continuous-time transfer function F at s.

F(omega,true) evaluates the discrete-time transfer function F at exp(imTsomega)

F(z,false) evaluates the discrete-time transfer function F at z

Base.step — Method

y, t, x = step(sys[, tfinal])
y, t, x = step(sys[, t])

Calculate the step response of system sys. If the final time tfinal or time vector t is not provided, one is calculated based on the system pole locations.

y has size (ny, length(t), nu), x has size (nx, length(t), nu)

CommonSolve.solve — Function

sol = solve(s::AbstractSimulator, x0, tspan,  args...; kwargs...)

Simulate the system represented by s from initial state x0 over time span tspan = (t0,tf). args and kwargs are sent to the solve function from OrdinaryDiffEq, e.g., solve(s, x0, tspan, Tsit5(), reltol=1e-5) solves the problem with solver Tsit5() and relative tolerance 1e-5.

See also Simulator lsim

ControlSystems.append — Method

append(systems::StateSpace...), append(systems::TransferFunction...)

Append systems in block diagonal form

ControlSystems.balance — Function

S, P, B = balance(A[, perm=true])

Compute a similarity transform T = S*P resulting in B = T\A*T such that the row and column norms of B are approximately equivalent. If perm=false, the transformation will only scale A using diagonal S, and not permute A (i.e., set P=I).

ControlSystems.balance_statespace — Function

A, B, C, T = balance_statespace{S}(A::Matrix{S}, B::Matrix{S}, C::Matrix{S}, perm::Bool=false)
sys, T = balance_statespace(sys::StateSpace, perm::Bool=false)

Computes a balancing transformation T that attempts to scale the system so that the row and column norms of [TA/T TB; C/T 0] are approximately equal. If perm=true, the states in A are allowed to be reordered.

The inverse of sysb, T = balance_statespace(sys) is given by similarity_transform(sysb, T)

This is not the same as finding a balanced realization with equal and diagonal observability and reachability gramians, see balreal

ControlSystems.balreal — Method

sysr, G, T = balreal(sys::StateSpace)

Calculates a balanced realization of the system sys, such that the observability and reachability gramians of the balanced system are equal and diagonal G. T is the similarity transform between the old state x and the new state z such that Tz = x.

See also gram, baltrunc

Reference: Varga A., Balancing-free square-root algorithm for computing singular perturbation approximations.

ControlSystems.baltrunc — Method

sysr, G, T = baltrunc(sys::StateSpace; atol = √ϵ, rtol=1e-3, n = nothing, residual = false)

Reduces the state dimension by calculating a balanced realization of the system sys, such that the observability and reachability gramians of the balanced system are equal and diagonal G, and truncating it to order n. If n is not provided, it's chosen such that all states corresponding to singular values less than atol and less that rtol σmax are removed.

T is the similarity transform between the old state x and the newstate z such that Tz = x.

If residual = true, matched static gain is achieved through "residualization", i.e., setting

\[0 = A_{21}x_{1} + A_{22}x_{2} + B_{2}u\]

where indices 1/2 correspond to the remaining/truncated states respectively.

See also gram, balreal

Glad, Ljung, Reglerteori: Flervariabla och Olinjära metoder

ControlSystems.bode — Method

mag, phase, w = bode(sys[, w])

Compute the magnitude and phase parts of the frequency response of system sys at frequencies w

mag and phase has size (length(w), ny, nu)

ControlSystems.bodeplot — Function

fig = bodeplot(sys, args...)
bodeplot(LTISystem[sys1, sys2...], args...; plotphase=true, kwargs...)

Create a Bode plot of the LTISystem(s). A frequency vector w can be optionally provided. To change the Magnitude scale see setPlotScale(str)

If hz=true, the plot x-axis will be displayed in Hertz, the input frequency vector is still treated as rad/s.

kwargs is sent as argument to RecipesBase.plot.

ControlSystems.bodev — Method

bodev(sys::LTISystem, w::AbstractVector; )

For use with SISO systems where it acts the same as bode but with the extra dimensions removed in the returned values.

ControlSystems.bodev — Method

bodev(sys::LTISystem; )

For use with SISO systems where it acts the same as bode but with the extra dimensions removed in the returned values.

ControlSystems.c2d — Function

sysd = c2d(sys::AbstractStateSpace{<:Continuous}, Ts, method=:zoh; w_prewarp=0)
Gd = c2d(G::TransferFunction{<:Continuous}, Ts, method=:zoh)

Convert the continuous-time system sys into a discrete-time system with sample time Ts, using the specified method (:zoh, :foh, :fwdeuler or :tustin). Note that the forward-Euler method generally requires the sample time to be very small relative to the time constants of the system.

method = :tustin performs a bilinear transform with prewarp frequency w_prewarp.

w_prewarp: Frequency (rad/s) for pre-warping when usingthe Tustin method, has no effect for other methods.

See also c2d_x0map

ControlSystems.c2d — Function

c2d(G::DelayLtiSystem, Ts, method=:zoh)

ControlSystems.c2d_poly2poly — Method

c2d_poly2poly(ro, Ts)

returns the polynomial coefficients in discrete time given polynomial coefficients in continuous time

ControlSystems.c2d_roots2poly — Method

c2d_roots2poly(ro, Ts)

returns the polynomial coefficients in discrete time given a vector of roots in continuous time

ControlSystems.c2d_x0map — Function

sysd, x0map = c2d_x0map(sys::AbstractStateSpace{<:Continuous}, Ts, method=:zoh; w_prewarp=0)

Returns the discretization sysd of the system sys and a matrix x0map that transforms the initial conditions to the discrete domain by x0_discrete = x0map*[x0; u0]

See c2d for further details.

ControlSystems.care — Method

care(A, B, Q, R)

Compute 'X', the solution to the continuous-time algebraic Riccati equation, defined as A'X + XA - (XB)R^-1(B'X) + Q = 0, where R is non-singular.

Uses MatrixEquations.arec.

ControlSystems.covar — Method

P = covar(sys, W)

Calculate the stationary covariance P = E[y(t)y(t)'] of the output y of a StateSpace model sys driven by white Gaussian noise w with covariance E[w(t)w(τ)]=W*δ(t-τ) (δ is the Dirac delta).

Remark: If sys is unstable then the resulting covariance is a matrix of Infs. Entries corresponding to direct feedthrough (DWD' .!= 0) will equal Inf for continuous-time systems.

ControlSystems.ctrb — Method

ctrb(A, B) or ctrb(sys)

Compute the controllability matrix for the system described by (A, B) or sys.

Note that checking for controllability by computing the rank from ctrb is not the most numerically accurate way, a better method is checking if gram(sys, :c) is positive definite.

ControlSystems.d2c — Function

d2c(sys::AbstractStateSpace{<:Discrete}, method::Symbol = :zoh; w_prewarp=0)

Convert discrete-time system to a continuous time system, assuming that the discrete-time system was discretized using method. Available methods are `:zoh, :fwdeuler´.

w_prewarp: Frequency for pre-warping when usingthe Tustin method, has no effect for other methods.

ControlSystems.dab — Method

X,Y = dab(A,B,C)

DAB Solves the Diophantine-Aryabhatta-Bezout identity

AX + BY = C, where A, B, C, X and Y are polynomials and deg Y = deg A - 1.

See Computer-Controlled Systems: Theory and Design, Third Edition Karl Johan Åström, Björn Wittenmark

ControlSystems.damp — Method

Wn, zeta, ps = damp(sys)

Compute the natural frequencies, Wn, and damping ratios, zeta, of the poles, ps, of sys

ControlSystems.dampreport — Method

dampreport(sys)

Display a report of the poles, damping ratio, natural frequency, and time constant of the system sys

ControlSystems.dare — Method

dare(A, B, Q, R; kwargs...)

Compute X, the solution to the discrete-time algebraic Riccati equation, defined as A'XA - X - (A'XB)(B'XB + R)^-1(B'XA) + Q = 0, where Q>=0 and R>0

Uses MatrixEquations.ared. For keyword arguments, see the docstring of ControlSystems.MatrixEquations.ared

ControlSystems.dcgain — Function

dcgain(sys, ϵ=0)

Compute the dcgain of system sys.

equal to G(0) for continuous-time systems and G(1) for discrete-time systems.

ϵ can be provided to evaluate the dcgain with a small perturbation into the stability region of the complex plane.

ControlSystems.delaymargin — Method

dₘ = delaymargin(G::LTISystem)

Only supports SISO systems

ControlSystems.diagonalize — Method

dsys = diagonalize(s::StateSpace, digits=12) Diagonalizes the system such that the A-matrix is diagonal.

ControlSystems.dkalman — Method

dkalman(A, C, R1, R2)
dkalman(sys, R1, R2)

Calculate the optimal Kalman gain for discrete time systems

The args...; kwargs... are sent to the Riccati solver, allowing specification of cross-covariance etc. See ?MatrixEquations.ared for more help.

ControlSystems.dlqr — Method

dlqr(A, B, Q, R, args...; kwargs...)
dlqr(sys, Q, R, args...; kwargs...)

Calculate the optimal gain matrix K for the state-feedback law u[k] = -K*x[k] that minimizes the cost function:

J = sum(x'Qx + u'Ru, 0, inf).

For the discrte time model x[k+1] = Ax[k] + Bu[k].

See also lqg

The args...; kwargs... are sent to the Riccati solver, allowing specification of cross-covariance etc. See ?MatrixEquations.ared for more help.

Usage example:

using LinearAlgebra # For identity matrix I
using Plots
Ts = 0.1
A = [1 Ts; 0 1]
B = [0;1]
C = [1 0]
sys = ss(A,B,C,0, Ts)
Q = I
R = I
L = dlqr(A,B,Q,R) # lqr(sys,Q,R) can also be used

u(x,t) = -L*x # Form control law,
t=0:Ts:5
x0 = [1,0]
y, t, x, uout = lsim(sys,u,t,x0=x0)
plot(t,x', lab=["Position"  "Velocity"], xlabel="Time [s]")

ControlSystems.dlyap — Method

dlyap(A, Q; kwargs...)

Compute the solution X to the discrete Lyapunov equation AXA' - X + Q = 0.

Uses MatrixEquations.lyapd. For keyword arguments, see the docstring of ControlSystems.MatrixEquations.lyapd

ControlSystems.evalfr — Method

evalfr(sys, x) Evaluate the transfer function of the LTI system sys at the complex number s=x (continuous-time) or z=x (discrete-time).

For many values of x, use freqresp instead.

ControlSystems.feedback — Method

feedback(sys1::AbstractStateSpace, sys2::AbstractStateSpace;
         U1=:, Y1=:, U2=:, Y2=:, W1=:, Z1=:, W2=Int[], Z2=Int[],
         Wperm=:, Zperm=:, pos_feedback::Bool=false)

Basic use feedback(sys1, sys2) forms the feedback interconnection

           ┌──────────────┐
◄──────────┤     sys1     │◄──── Σ ◄──────
    │      │              │      │
    │      └──────────────┘      -1
    │                            |
    │      ┌──────────────┐      │
    └─────►│     sys2     ├──────┘
           │              │
           └──────────────┘

Advanced use feedback also supports more flexible use according to the figure below

              ┌──────────────┐
      z1◄─────┤     sys1     │◄──────w1
 ┌─── y1◄─────┤              │◄──────u1 ◄─┐
 │            └──────────────┘            │
 │                                        α
 │            ┌──────────────┐            │
 └──► u2─────►│     sys2     ├───────►y2──┘
      w2─────►│              ├───────►z2
              └──────────────┘

U1, W1 specifies the indices of the input signals of sys1 corresponding to u1 and w1 Y1, Z1 specifies the indices of the output signals of sys1 corresponding to y1 and z1 U2, W2, Y2, Z2 specifies the corresponding signals of sys2

Specify Wperm and Zperm to reorder the inputs (corresponding to [w1; w2]) and outputs (corresponding to [z1; z2]) in the resulting statespace model.

Negative feedback (α = -1) is the default. Specify pos_feedback=true for positive feedback (α = 1).

See also lft, starprod.

See Zhou, Doyle, Glover (1996) for similar (somewhat less symmetric) formulas.

ControlSystems.feedback — Method

feedback(sys)
feedback(sys1, sys2)

For a general LTI-system, feedback forms the negative feedback interconnection

>-+ sys1 +-->
  |      |
 (-)sys2 +

If no second system is given, negative identity feedback is assumed

ControlSystems.feedback2dof — Method

feedback2dof(P,R,S,T)
feedback2dof(B,A,R,S,T)

Return BT/(AR+ST) where B and A are the numerator and denomenator polynomials of P respectively

Return BT/(AR+ST)

ControlSystems.feedback2dof — Method

feedback2dof(P::TransferFunction, C::TransferFunction, F::TransferFunction)

Return the transfer function P(F+C)/(1+PC) which is the closed-loop system with process P, controller C and feedforward filter F from reference to control signal (by-passing C).

         +-------+
         |       |
   +----->   F   +----+
   |     |       |    |
   |     +-------+    |
   |     +-------+    |    +-------+
r  |  -  |       |    |    |       |    y
+--+----->   C   +----+---->   P   +---+-->
      |  |       |         |       |   |
      |  +-------+         +-------+   |
      |                                |
      +--------------------------------+

ControlSystems.freqresp — Method

sys_fr = freqresp(sys, w)

Evaluate the frequency response of a linear system

w -> C*((iw*im -A)^-1)*B + D

of system sys over the frequency vector w.

ControlSystems.freqrespv — Method

freqrespv(sys::LTISystem, w_vec::AbstractVector{<:Real}; )

For use with SISO systems where it acts the same as freqresp but with the extra dimensions removed in the returned values.

ControlSystems.freqrespv — Method

freqrespv(sys::AbstractStateSpace, w_vec::AbstractVector{W}; )

For use with SISO systems where it acts the same as freqresp but with the extra dimensions removed in the returned values.

ControlSystems.gangoffour — Method

S, PS, CS, T = gangoffour(P, C; minimal=true)
gangoffour(P::AbstractVector, C::AbstractVector; minimal=true)

Given a transfer function describing the plant P and a transfer function describing the controller C, computes the four transfer functions in the Gang-of-Four.

S = 1/(1+PC) Sensitivity function

PS = P/(1+PC) Load disturbance to measurement signal

CS = C/(1+PC) Measurement noise to control signal

T = PC/(1+PC) Complementary sensitivity function

Only supports SISO systems

ControlSystems.gangoffourplot — Method

fig = gangoffourplot(P::LTISystem, C::LTISystem; minimal=true, plotphase=false, kwargs...)
gangoffourplot(P::Union{Vector, LTISystem}, C::Vector; minimal=true, plotphase=false, kwargs...)

Gang-of-Four plot. kwargs is sent as argument to RecipesBase.plot.

ControlSystems.gangofseven — Method

S, PS, CS, T, RY, RU, RE = gangofseven(P,C,F)

Given transfer functions describing the Plant P, the controller C and a feed forward block F, computes the four transfer functions in the Gang-of-Four and the transferfunctions corresponding to the feed forward.

S = 1/(1+PC) Sensitivity function

PS = P/(1+PC)

CS = C/(1+PC)

T = PC/(1+PC) Complementary sensitivity function

RY = PCF/(1+PC)

RU = CF/(1+P*C)

RE = F/(1+P*C)

Only supports SISO systems

ControlSystems.gram — Method

gram(sys, opt; kwargs...)

Compute the grammian of system sys. If opt is :c, computes the controllability grammian. If opt is :o, computes the observability grammian.

See also grampd For keyword arguments, see grampd.

ControlSystems.grampd — Method

U = grampd(sys, opt; kwargs...)

Return a Cholesky factor U of the grammian of system sys. If opt is :c, computes the controllability grammian G = U*U'. If opt is :o, computes the observability grammian G = U'U.

Obtain a Cholesky object by Cholesky(U) for observability grammian

Uses MatrixEquations.plyapc/plyapd. For keyword arguments, see the docstring of ControlSystems.MatrixEquations.plyapc/plyapd

ControlSystems.hinfnorm — Method

(Ninf, ω_peak) = hinfnorm(sys; tol=1e-6)

Compute the H∞ norm Ninf of the LTI system sys, together with a frequency ω_peak at which the gain Ninf is achieved.

Ninf := sup_ω σ_max[sys(iω)] if G is stable (σ_max = largest singular value) := Inf' ifG` is unstable

tol is an optional keyword argument for the desired relative accuracy for the computed H∞ norm (not an absolute certificate).

sys is first converted to a state space model if needed.

The continuous-time L∞ norm computation implements the 'two-step algorithm' in:
N.A. Bruinsma and M. Steinbuch, 'A fast algorithm to compute the H∞-norm of a transfer function matrix', Systems and Control Letters (1990), pp. 287-293.

For the discrete-time version, see:
P. Bongers, O. Bosgra, M. Steinbuch, 'L∞-norm calculation for generalized state space systems in continuous and discrete time', American Control Conference, 1991.

See also linfnorm.

ControlSystems.impulse — Method

y, t, x = impulse(sys[, tfinal])
y, t, x = impulse(sys[, t])

Calculate the impulse response of system sys. If the final time tfinal or time vector t is not provided, one is calculated based on the system pole locations.

y has size (ny, length(t), nu), x has size (nx, length(t), nu)

ControlSystems.innovation_form — Method

sysi = innovation_form(sys, R1, R2)
sysi = innovation_form(sys; sysw=I, syse=I, R1=I, R2=I)

Takes a system

x' = Ax + Bu + w ~ R1
y  = Cx + e ~ R2

and returns the system

x' = Ax + Kv
y  = Cx + v

where v is the innovation sequence.

If sysw (syse) is given, the covariance resulting in filtering noise with R1 (R2) through sysw (syse) is used as covariance.

See Stochastic Control, Chapter 4, Åström

ControlSystems.iscontinuous — Method

iscontinuous(sys)

Returns true for a continuous-time system sys, else returns false.

ControlSystems.isdiscrete — Method

isdiscrete(sys)

Returns true for a discrete-time system sys, else returns false.

ControlSystems.isproper — Method

isproper(tf)

Returns true if the TransferFunction is proper. This means that order(den)

= order(num))

ControlSystems.isstable — Method

isstable(sys)

Returns true if sys is stable, else returns false.

ControlSystems.kalman — Method

kalman(A, C, R1, R2)
kalman(sys, R1, R2)

Calculate the optimal Kalman gain

The args...; kwargs... are sent to the Riccati solver, allowing specification of cross-covariance etc. See ?MatrixEquations.arec/ared for more help.

See also LQG

ControlSystems.laglink — Method

laglink(a, M; Ts=0)

Returns a phase retarding link, the rule of thumb a = 0.1ωc guarantees less than 6 degrees phase margin loss. The bode curve will go from M, bend down at a/M and level out at 1 for frequencies > a

ControlSystems.leadlink — Method

leadlink(b, N, K; Ts=0)

Returns a phase advancing link, the top of the phase curve is located at ω = b√(N) where the link amplification is K√(N) The bode curve will go from K, bend up at b and level out at KN for frequencies > bN

The phase advance at ω = b√(N) can be plotted as a function of N with leadlinkcurve()

Values of N < 1 will give a phase retarding link.

See also leadlinkat laglink

ControlSystems.leadlinkat — Method

leadlinkat(ω, N, K; Ts=0)

Returns a phase advancing link, the top of the phase curve is located at ω where the link amplification is K√(N) The bode curve will go from K, bend up at ω/√(N) and level out at KN for frequencies > ω√(N)

The phase advance at ω can be plotted as a function of N with leadlinkcurve()

Values of N < 1 will give a phase retarding link.

See also leadlink laglink

ControlSystems.leadlinkcurve — Function

leadlinkcurve(start=1)

Plot the phase advance as a function of N for a lead link (phase advance link) If an input argument s is given, the curve is plotted from s to 10, else from 1 to 10.

See also Leadlink, leadlinkat

ControlSystems.lft — Function

lft(G, Δ, type=:l)

Lower and upper linear fractional transformation between systems G and Δ.

Specify :l lor lower LFT, and :u for upper LFT.

G must have more inputs and outputs than Δ has outputs and inputs.

For details, see Chapter 9.1 in Zhou, K. and JC Doyle. Essentials of robust control, Prentice hall (NJ), 1998

ControlSystems.linfnorm — Method

(Ninf, ω_peak) = linfnorm(sys; tol=1e-6)

Compute the L∞ norm Ninf of the LTI system sys, together with a frequency ω_peak at which the gain Ninf is achieved.

Ninf := sup_ω σ_max[sys(iω)] (σ_max denotes the largest singular value)

tol is an optional keyword argument representing the desired relative accuracy for the computed L∞ norm (this is not an absolute certificate however).

sys is first converted to a state space model if needed.

The continuous-time L∞ norm computation implements the 'two-step algorithm' in:
N.A. Bruinsma and M. Steinbuch, 'A fast algorithm to compute the H∞-norm of a transfer function matrix', Systems and Control Letters (1990), pp. 287-293.

For the discrete-time version, see:
P. Bongers, O. Bosgra, M. Steinbuch, 'L∞-norm calculation for generalized state space systems in continuous and discrete time', American Control Conference, 1991.

See also hinfnorm.

ControlSystems.loopshapingPI — Method

kp,ki,C = loopshapingPI(P,ωp; ϕl,rl, phasemargin, doplot = false)

Selects the parameters of a PI-controller such that the Nyquist curve of P at the frequency ωp is moved to rl exp(i ϕl)

If phasemargin is supplied, ϕl is selected such that the curve is moved to an angle of phasemargin - 180 degrees

If no rl is given, the magnitude of the curve at ωp is kept the same and only the phase is affected, the same goes for ϕl if no phasemargin is given.

Set doplot = true to plot the gangoffourplot and nyquistplot of the system.

See also pidplots, stabregionPID

ControlSystems.lqr — Method

lqr(A, B, Q, R, args...; kwargs...)

Calculate the optimal gain matrix K for the state-feedback law u = -K*x that minimizes the cost function:

J = integral(x'Qx + u'Ru, 0, inf).

For the continuous time model dx = Ax + Bu.

lqr(sys, Q, R)

Solve the LQR problem for state-space system sys. Works for both discrete and continuous time systems.

The args...; kwargs... are sent to the Riccati solver, allowing specification of cross-covariance etc. See ?MatrixEquations.arec for more help.

See also LQG Usage example:

using LinearAlgebra # For identity matrix I
using Plots
A = [0 1; 0 0]
B = [0;1]
C = [1 0]
sys = ss(A,B,C,0)
Q = I
R = I
L = lqr(sys,Q,R)

u(x,t) = -L*x # Form control law,
t=0:0.1:5
x0 = [1,0]
y, t, x, uout = lsim(sys,u,t,x0=x0)
plot(t,x', lab=["Position" "Velocity"], xlabel="Time [s]")

ControlSystems.lsim — Method

result = lsim(sys, u[, t]; x0, method])
result = lsim(sys, u::Function, t; x0, method)

Calculate the time response of system sys to input u. If x0 is ommitted, a zero vector is used.

The result structure contains the fields y, t, x, u and can be destructured automatically by iteration, e.g.,

y, t, x, u = result

y, x, u have time in the second dimension. Initial state x0 defaults to zero.

Continuous time systems are simulated using an ODE solver if u is a function. If u is an array, the system is discretized (with method=:zoh by default) before simulation. For a lower level inteface, see ?Simulator and ?solve

u can be a function or a matrix/vector of precalculated control signals. If u is a function, then u(x,i) (u(x,t)) is called to calculate the control signal every iteration (time instance used by solver). This can be used to provide a control law such as state feedback u(x,t) = -L*x calculated by lqr. To simulate a unit step, use (x,i)-> 1, for a ramp, use (x,i)-> i*Ts, for a step at t=5, use (x,i)-> (i*Ts >= 5) etc.

Usage example:

using LinearAlgebra # For identity matrix I
using Plots

A = [0 1; 0 0]
B = [0;1]
C = [1 0]
sys = ss(A,B,C,0)
Q = I
R = I
L = lqr(sys,Q,R)

u(x,t) = -L*x # Form control law,
t=0:0.1:5
x0 = [1,0]
y, t, x, uout = lsim(sys,u,t,x0=x0)
plot(t,x', lab=["Position" "Velocity"], xlabel="Time [s]")

ControlSystems.lsim — Method

`y, t, x = lsim(sys::DelayLtiSystem, u, t::AbstractArray{<:Real}; x0=fill(0.0, nstates(sys)), alg=MethodOfSteps(Tsit5()), abstol=1e-6, reltol=1e-6, kwargs...)`

Simulate system `sys`, over time `t`, using input signal `u`, with initial state `x0`, using method `alg` .

Arguments:

`t`: Has to be an `AbstractVector` with equidistant time samples (`t[i] - t[i-1]` constant)
`u`: Function to determine control signal `ut` at a time `t`, on any of the following forms:
    Can be a constant `Number` or `Vector`, interpreted as `ut .= u` , or
    Function `ut .= u(t)`, or
    In-place function `u(ut, t)`. (Slightly more effienct)
`alg, abstol, reltol` and `kwargs...`: are sent to `DelayDiffEq.solve`.

Returns: times `t`, and `y` and `x` at those times.

ControlSystems.margin — Method

ωgₘ, gₘ, ωϕₘ, ϕₘ = margin{S<:Real}(sys::LTISystem, w::AbstractVector{S}; full=false, allMargins=false)

returns frequencies for gain margins, gain margins, frequencies for phase margins, phase margins

If !allMargins, return only the smallest margin

If full return also fullPhase

ControlSystems.markovparam — Method

markovparam(sys, n)

Compute the nth markov parameter of state-space system sys. This is defined as the following:

h(0) = D

h(n) = C*A^(n-1)*B

ControlSystems.minreal — Function

tf = minreal(tf::TransferFunction, eps=sqrt(eps()))

Create a minimial representation of each transfer function in tf by cancelling poles and zeros will promote system to an appropriate numeric type

ControlSystems.minreal — Method

minreal(sys::T; fast=false, kwargs...)

Minimal realisation algorithm from P. Van Dooreen, The generalized eigenstructure problem in linear system theory, IEEE Transactions on Automatic Control

For information about the options, see ?ControlSystems.MatrixPencils.lsminreal

ControlSystems.nicholsplot — Function

fig = nicholsplot{T<:LTISystem}(systems::Vector{T}, w::AbstractVector; kwargs...)

Create a Nichols plot of the LTISystem(s). A frequency vector w can be optionally provided.

Keyword arguments:

text = true
Gains = [12, 6, 3, 1, 0.5, -0.5, -1, -3, -6, -10, -20, -40, -60]
pInc = 30
sat = 0.4
val = 0.85
fontsize = 10

pInc determines the increment in degrees between phase lines.

sat ∈ [0,1] determines the saturation of the gain lines

val ∈ [0,1] determines the brightness of the gain lines

Additional keyword arguments are sent to the function plotting the systems and can be used to specify colors, line styles etc. using regular RecipesBase.jl syntax

This function is based on code subject to the two-clause BSD licence Copyright 2011 Will Robertson Copyright 2011 Philipp Allgeuer

ControlSystems.nyquist — Method

re, im, w = nyquist(sys[, w])

Compute the real and imaginary parts of the frequency response of system sys at frequencies w

re and im has size (length(w), ny, nu)

ControlSystems.nyquistplot — Function

fig = nyquistplot(sys; Ms_circles=Float64[], unit_circle=false, hz = false, kwargs...)
nyquistplot(LTISystem[sys1, sys2...]; Ms_circles=Float64[], unit_circle=false, kwargs...)

Create a Nyquist plot of the LTISystem(s). A frequency vector w can be optionally provided.

unit_circle: if the unit circle should be displayed

Ms_circles: draw circles corresponding to given levels of sensitivity (circles around -1 with radii 1/Ms). Ms_circles can be supplied as a number or a vector of numbers. A design staying outside such a circle has a phase margin of at least 2asin(1/(2Ms)) rad and a gain margin of at least Ms/(Ms-1).

If hz=true, the hover information will be displayed in Hertz, the input frequency vector is still treated as rad/s.

kwargs is sent as argument to plot.

ControlSystems.nyquistv — Method

nyquistv(sys::LTISystem, w::AbstractVector; )

For use with SISO systems where it acts the same as nyquist but with the extra dimensions removed in the returned values.

ControlSystems.nyquistv — Method

nyquistv(sys::LTISystem; )

For use with SISO systems where it acts the same as nyquist but with the extra dimensions removed in the returned values.

ControlSystems.observer_controller — Method

cont = observer_controller(sys, L::AbstractMatrix, K::AbstractMatrix)

Return the observer_controller cont that is given by ss(A - B*L - K*C + K*D*L, K, L, 0)

Such that feedback(sys, cont) produces a closed-loop system with eigenvalues given by A-KC and A-BL.

Arguments:

sys: Model of system

L: State-feedback gain u = -Lx

K: Observer gain

ControlSystems.observer_predictor — Method

observer_predictor(sys::AbstractStateSpace, R1, R2)
observer_predictor(sys::AbstractStateSpace, K)

Return the observer_predictor system x̂' = (A - KC)x̂ + (B-KD)u + Ky ŷ = Cx + Du with the input equation [B K] * [u; y]

If covariance matrices R1, R2 are given, the kalman gain K is calculaded.

See also innovation_form.

ControlSystems.obsv — Function

obsv(A, C, n=size(A,1))
obsv(sys, n=sys.nx)

Compute the observability matrix with n rows for the system described by (A, C) or sys. Providing the optional n > sys.nx returns an extended observability matrix.

Note that checking for observability by computing the rank from obsv is not the most numerically accurate way, a better method is checking if gram(sys, :o) is positive definite.

ControlSystems.pade — Method

pade(G::DelayLtiSystem, N)

Approximate all time-delays in G by Padé approximations of degree N.

ControlSystems.pade — Method

pade(τ::Real, N::Int)

Compute the Nth order Padé approximation of a time-delay of length τ.

ControlSystems.parallel — Method

parallel(sys1::LTISystem, sys2::LTISystem)

Connect systems in parallel, equivalent to sys2+sys1

ControlSystems.pid — Method

C = pid(; kp=0, ki=0; kd=0, time=false, series=false)

Calculates and returns a PID controller on transfer function form.

time indicates whether or not the parameters are given as gains (default) or as time constants

series indicates whether or not the series form or parallel form (default) is desired

ControlSystems.pidplots — Method

pidplots(P, args...; kps=0, kis=0, kds=0, time=false, series=false, ω=0)

Plots interesting figures related to closing the loop around process P with a PID controller Send in a bunch of PID-parameters in any of the vectors kp, ki, kd. The vectors must be the same length.

-time indicates whether or not the parameters are given as gains (default) or as time constants -series indicates whether or not the series form or parallel form (default) is desired

Available plots are :gof for Gang of four, :nyquist, :controller for a bode plot of the controller TF and :pz for pole-zero maps

One can also supply a frequency vector ω to be used in Bode and Nyquist plots

See also loopshapingPI, stabregionPID

ControlSystems.place — Function

place(A, B, p, opt=:c)
place(sys::StateSpace, p, opt=:c)

Calculate the gain matrix K such that A - BK has eigenvalues p.

place(A, C, p, opt=:o)
place(sys::StateSpace, p, opt=:o)

Calculate the observer gain matrix L such that A - LC has eigenvalues p.

Uses Ackermann's formula. Currently handles only SISO systems.

ControlSystems.placePI — Method

piparams, C = placePI(P, ω₀, ζ; form=:standard)

Selects the parameters of a PI-controller such that the poles of closed loop between P and C are placed to match the poles of s^2 + 2ζω₀s + ω₀^2.

The form keyword allows you to choose which form the PI parameters should be returned on.

:standard - Kp*(1 + 1/Ti/s)

:series - Kc*(1 + 1/τi/s)

:parallel - Kp + Ki/s

:Ti - Kp + 1/(s*Ti) (non-standard form sometimes used in industry)

piparams is a named tuple with the controller parameters.

C is the transfer function of the controller.

ControlSystems.poles — Method

poles(sys)

Compute the poles of system sys.

ControlSystems.prescale — Method

syst, S = prescale(sys)

Perform a eigendecomposition on system state-transition matrix sys.A.

Ã = S⁻¹AS
B̃ = S⁻¹ B
C̃ = CS
D̃ = D

Such that Ã is diagonal. Returns a new scaled state-space object and the associated transformation matrix.

ControlSystems.pzmap — Function

fig = pzmap(fig, system, args...; kwargs...)

Create a pole-zero map of the LTISystem(s) in figure fig, args and kwargs will be sent to the scatter plot command.

ControlSystems.reduce_sys — Method

Implements REDUCE in the Emami-Naeini & Van Dooren paper. Returns transformed A, B, C, D matrices. These are empty if there are no zeros.

ControlSystems.relative_gain_array — Method

relative_gain_array(A::AbstractMatrix; tol = 1.0e-15)

Reference: "On the Relative Gain Array (RGA) with Singular and Rectangular Matrices" Jeffrey Uhlmann https://arxiv.org/pdf/1805.10312.pdf

ControlSystems.relative_gain_array — Method

relative_gain_array(G, w::AbstractVector)
relative_gain_array(G, w::Number)

Calculate the relative gain array of G at frequencies w. G(iω) .* pinv(tranpose(G(iω)))

The RGA can be used to find input-output pairings for MIMO control using individially tuned loops. Pair the inputs and outputs such that the RGA(ωc) at the crossover frequency becomes as close to diagonal as possible. Avoid pairings such that RGA(0) contains negative diagonal elements.

The sum of the absolute values of the entries in the RGA is a good measure of the "true condition number" of G, the best condition number that can be achieved by input/output scaling of G, -Glad, Ljung.

The RGA is invariant to input/output scaling of G.

If the RGA contains large entries, the system may be sensitive to model errors, -Skogestad, "Multivariable Feedback Control: Analysis and Design":

Uncertainty in the input channels (diagonal input uncertainty). Plants with

large RGA-elements around the crossover frequency are fundamentally difficult to control because of sensitivity to input uncertainty (e.g. caused by uncertain or neglected actuator dynamics). In particular, decouplers or other inverse-based controllers should not be used for plants with large RGAeleme

Element uncertainty. Large RGA-elements imply sensitivity to element-by-element uncertainty.

However, this kind of uncertainty may not occur in practice due to physical couplings between the transfer function elements. Therefore, diagonal input uncertainty (which is always present) is usually of more concern for plants with large RGAelemen

The relative gain array is computed using the The unit-consistent (UC) generalized inverse Reference: "On the Relative Gain Array (RGA) with Singular and Rectangular Matrices" Jeffrey Uhlmann https://arxiv.org/pdf/1805.10312.pdf

ControlSystems.rgaplot — Function

rgaplot(sys, args...; hz=false)
rgaplot(LTISystem[sys1, sys2...], args...; hz=false)

Plot the relative-gain array entries of the LTISystem(s). A frequency vector w can be optionally provided.

If hz=true, the plot x-axis will be displayed in Hertz, the input frequency vector is still treated as rad/s.

kwargs is sent as argument to Plots.plot.

ControlSystems.rlocusplot — Function

rlocusplot(P::LTISystem, K)

Computes and plots the root locus of the SISO LTISystem P with a negative feedback loop and feedback gains K, if K is not provided, range(1e-6,stop=500,length=10000) is used. If OrdinaryDiffEq.jl is installed and loaded by the user (using OrdinaryDiffEq), rlocusplot will use an adaptive step-size algorithm to select values of K. A scalar Kmax can then be given as second argument.

ControlSystems.rstc — Method

See ?rstd for the discerte case

ControlSystems.rstd — Method

R,S,T=rstd(BPLUS,BMINUS,A,BM1,AM,AO,AR,AS)
R,S,T=rstd(BPLUS,BMINUS,A,BM1,AM,AO,AR)
R,S,T=rstd(BPLUS,BMINUS,A,BM1,AM,AO)

rstd Polynomial synthesis in discrete time.

Polynomial synthesis according to CCS ch 10 to design a controller R(q) u(k) = T(q) r(k) - S(q) y(k)

Inputs: BPLUS : Part of open loop numerator BMINUS : Part of open loop numerator A : Open loop denominator BM1 : Additional zeros AM : Closed loop denominator AO : Observer polynomial AR : Pre-specified factor of R, e.g integral part [1, -1]^k AS : Pre-specified factor of S, e.g notch filter [1, 0, w^2]

Outputs: R,S,T : Polynomials in controller

See function dab how the solution to the Diophantine- Aryabhatta-Bezout identity is chosen.

See Computer-Controlled Systems: Theory and Design, Third Edition Karl Johan Åström, Björn Wittenmark

ControlSystems.series — Method

series(sys1::LTISystem, sys2::LTISystem)

Connect systems in series, equivalent to sys2*sys1

ControlSystems.setPlotScale — Method

setPlotScale(str)

Set the default scale of magnitude in bodeplot and sigmaplot. str should be either "dB" or "log10".

ControlSystems.sigma — Method

sv, w = sigma(sys[, w])

Compute the singular values sv of the frequency response of system sys at frequencies w

sv has size (length(w), max(ny, nu))

ControlSystems.sigmaplot — Function

sigmaplot(sys, args...; hz=false)
sigmaplot(LTISystem[sys1, sys2...], args...; hz=false)

Plot the singular values of the frequency response of the LTISystem(s). A frequency vector w can be optionally provided.

If hz=true, the plot x-axis will be displayed in Hertz, the input frequency vector is still treated as rad/s.

kwargs is sent as argument to Plots.plot.

ControlSystems.sigmav — Method

sigmav(sys::LTISystem, w::AbstractVector; )

For use with SISO systems where it acts the same as sigma but with the extra dimensions removed in the returned values.

ControlSystems.sigmav — Method

sigmav(sys::LTISystem; )

For use with SISO systems where it acts the same as sigma but with the extra dimensions removed in the returned values.

ControlSystems.similarity_transform — Method

syst = similarity_transform(sys, T)

Perform a similarity transform T : Tx̃ = x on sys such that

Ã = T⁻¹AT
B̃ = T⁻¹ B
C̃ = CT
D̃ = D

ControlSystems.sminreal — Method

sminreal(sys)

Compute the structurally minimal realization of the state-space system sys. A structurally minimal realization is one where only states that can be determined to be uncontrollable and unobservable based on the location of 0s in sys are removed.

Systems with numerical noise in the coefficients, e.g., noise on the order of eps require truncation to zero to be affected by structural simplification, e.g.,

trunc_zero!(A) = A[abs.(A) .< 10eps(maximum(abs, A))] .= 0
trunc_zero!(sys.A); trunc_zero!(sys.B); trunc_zero!(sys.C)
sminreal(sys)

ControlSystems.ss — Method

`sys = ss(A, B, C, D [,Ts])`

Create a state-space model sys::StateSpace{TE, T} with matrix element type T and TE is Continuous or <:Discrete.

This is a continuous-time model if Ts is omitted. Otherwise, this is a discrete-time model with sampling period Ts.

sys = ss(D [, Ts]) specifies a static gain matrix D.

ControlSystems.ssdata — Method

`A, B, C, D = ssdata(sys)`

ControlSystems.ssrand — Method

ssrand(T::Type, ny::Int, nu::Int, nstates::Int; proper=false, stable=true, Ts=nothing)

Returns a random StateSpace model with ny outputs, nu inputs, and nstates states, whose matrix elements are normally distributed.

It is possible to specify if the system should be proper or stable.

Specify a sample time Ts to obtain a discrete-time system.

ControlSystems.stabregionPID — Function

fig, kp, ki = stabregionPID(P, [ω]; kd=0, doplot = true)

Segments of the curve generated by this program is the boundary of the stability region for a process with transfer function P(s) The PID controller is assumed to be on the form kp +ki/s +kd s

The curve is found by analyzing P(s)*C(s) = -1 ⟹ |PC| = |P| |C| = 1 arg(P) + arg(C) = -π

If P is a function (e.g. s -> exp(-sqrt(s)) ), the stability of feedback loops using PI-controllers can be analyzed for processes with models with arbitrary analytic functions

See also stabregionPID, loopshapingPI, pidplots

ControlSystems.starprod — Method

starprod(sys1, sys2, dimu, dimy)

Compute the Redheffer star product.

length(U1) = length(Y2) = dimu and length(Y1) = length(U2) = dimy

For details, see Chapter 9.3 in Zhou, K. and JC Doyle. Essentials of robust control, Prentice hall (NJ), 1998

ControlSystems.tf — Method

sys = tf(num, den[, Ts]), sys = tf(gain[, Ts])

Create as a fraction of polynomials:

sys::TransferFunction{SisoRational{T,TR}} = numerator/denominator where T is the type of the coefficients in the polynomial.

num: the coefficients of the numerator polynomial. Either scalar or vector to create SISO systems or an array of vectors to create MIMO system.

den: the coefficients of the denominator polynomial. Either vector to create SISO systems or an array of vectors to create MIMO system.

Ts: Sample time if discrete time system.

Other uses: tf(sys): Convert sys to tf form. tf("s"), tf("z"): Create the continuous transferfunction s.

See also: zpk, ss

ControlSystems.tzeros — Method

tzeros(sys)

Compute the invariant zeros of the system sys. If sys is a minimal realization, these are also the transmission zeros.

ControlSystems.zpconv — Method

zpc(a,r,b,s)

form conv(a,r) + conv(b,s) where the lengths of the polynomials are equalized by zero-padding such that the addition can be carried out

ControlSystems.zpk — Method

zpk(gain[, Ts]), zpk(num, den, k[, Ts]), zpk(sys)

Create transfer function on zero pole gain form. The numerator and denominator are represented by their poles and zeros.

sys::TransferFunction{SisoZpk{T,TR}} = k*numerator/denominator where T is the type of k and TR the type of the zeros/poles, usually Float64 and Complex{Float64}.

num: the roots of the numerator polynomial. Either scalar or vector to create SISO systems or an array of vectors to create MIMO system.

den: the roots of the denominator polynomial. Either vector to create SISO systems or an array of vectors to create MIMO system.

k: The gain of the system. Obs, this is not the same as dcgain.

Ts: Sample time if discrete time system.

Other uses:

zpk(sys): Convert sys to zpk form.

zpk("s"): Create the transferfunction s.

ControlSystems.zpkdata — Method

z, p, k = zpkdata(sys)

Compute the zeros, poles, and gains of system sys.

Returns

z : Matrix{Vector{ComplexF64}}, (ny x nu)

p : Matrix{Vector{ComplexF64}}, (ny x nu)

k : Matrix{Float64}, (ny x nu)

LinearAlgebra.norm — Function

.. norm(sys, p=2; tol=1e-6)

norm(sys) or norm(sys,2) computes the H2 norm of the LTI system sys.

norm(sys, Inf) computes the H∞ norm of the LTI system sys. The H∞ norm is the same as the H∞ for stable systems, and Inf for unstable systems. If the peak gain frequency is required as well, use the function hinfnorm instead. See hinfnorm for further documentation.

tol is an optional keyword argument, used only for the computation of L∞ norms. It represents the desired relative accuracy for the computed L∞ norm (this is not an absolute certificate however).

sys is first converted to a StateSpace model if needed.