ControlSystems.bodeplot

ControlSystems.gangoffourplot

ControlSystems.impulseplot

ControlSystems.lsimplot

ControlSystems.marginplot

ControlSystems.nicholsplot

ControlSystems.nyquistplot

ControlSystems.pzmap

ControlSystems.setPlotScale

ControlSystems.sigmaplot

ControlSystems.stepplot

Plotting functions

ControlSystems.bodeplot — Function

fig = bodeplot(sys, args...), bodeplot(LTISystem[sys1, sys2...], args...; plotphase=true, kwargs...)

Create a Bode plot of the LTISystem(s). A frequency vector w can be optionally provided. To change the Magnitude scale see setPlotScale(str)

kwargs is sent as argument to Plots.plot.

ControlSystems.gangoffourplot — Method

fig = gangoffourplot(P::LTISystem, C::LTISystem), gangoffourplot(P::Union{Vector, LTISystem}, C::Vector; plotphase=false)

Gang-of-Four plot.

kwargs is sent as argument to Plots.plot.

ControlSystems.impulseplot — Function

`impulseplot(sys[, Tf]; kwargs...)`` or `impulseplot(sys[, t]; kwargs...)``

Plot impulse response of sys until final time Tf or at time points in the vector t. If not defined, suitable values are chosen based on sys. See also impulse

kwargs is sent as argument to Plots.plot.

ControlSystems.lsimplot — Function

fig = lsimplot(sys::LTISystem, u, t; x0=0, method)

lsimplot(LTISystem[sys1, sys2...], u, t; x0, method)

Calculate the time response of the LTISystem(s) to input u. If x0 is not specified, a zero vector is used.

Continuous time systems are discretized before simulation. By default, the method is chosen based on the smoothness of the input signal. Optionally, the method parameter can be specified as either :zoh or :foh.

ControlSystems.marginplot — Method

fig = marginplot(sys::LTISystem [,w::AbstractVector]; kwargs...), marginplot(sys::Vector{LTISystem}, w::AbstractVector; kwargs...)

Plot all the amplitude and phase margins of the system(s) sys. A frequency vector w can be optionally provided.

kwargs is sent as argument to Plots.plot.

ControlSystems.nicholsplot — Function

fig = nicholsplot{T<:LTISystem}(systems::Vector{T}, w::AbstractVector; kwargs...)

Create a Nichols plot of the LTISystem(s). A frequency vector w can be optionally provided.

Keyword arguments:

text = true
Gains = [12, 6, 3, 1, 0.5, -0.5, -1, -3, -6, -10, -20, -40, -60]
pInc = 30
sat = 0.4
val = 0.85
fontsize = 10

pInc determines the increment in degrees between phase lines.

sat ∈ [0,1] determines the saturation of the gain lines

val ∈ [0,1] determines the brightness of the gain lines

Additional keyword arguments are sent to the function plotting the systems and can be used to specify colors, line styles etc. using regular Plots.jl syntax

This function is based on code subject to the two-clause BSD licence Copyright 2011 Will Robertson Copyright 2011 Philipp Allgeuer

ControlSystems.nyquistplot — Function

fig = nyquistplot(sys; gaincircles=true, kwargs...), nyquistplot(LTISystem[sys1, sys2...]; gaincircles=true, kwargs...)

Create a Nyquist plot of the LTISystem(s). A frequency vector w can be optionally provided.

gaincircles plots the circles corresponding to |S(iω)| = 1 and |T(iω)| = 1, where S and T are the sensitivity and complementary sensitivity functions.

kwargs is sent as argument to plot.

ControlSystems.pzmap — Function

fig = pzmap(fig, system, args...; kwargs...)

Create a pole-zero map of the LTISystem(s) in figure fig, args and kwargs will be sent to the scatter plot command.

ControlSystems.setPlotScale — Method

setPlotScale(str)

Set the default scale of magnitude in bodeplot and sigmaplot. str should be either "dB" or "log10".

ControlSystems.sigmaplot — Function

sigmaplot(sys, args...), sigmaplot(LTISystem[sys1, sys2...], args...)

Plot the singular values of the frequency response of the LTISystem(s). A frequency vector w can be optionally provided.

kwargs is sent as argument to Plots.plot.

ControlSystems.stepplot — Function

stepeplot(sys[, Tf]; kwargs...)orstepplot(sys[, t]; kwargs...)`Plot step response ofsysuntil final timeTfor at time points in the vectort. If not defined, suitable values are chosen based onsys. See also [step`](@ref)

kwargs is sent as argument to Plots.plot.