ControlSystems.are

ControlSystems.are

ControlSystems.balance

ControlSystems.balreal

ControlSystems.baltrunc

ControlSystems.covar

ControlSystems.ctrb

ControlSystems.damp

ControlSystems.dampreport

ControlSystems.dcgain

ControlSystems.delaymargin

ControlSystems.gangoffour

ControlSystems.gangofseven

ControlSystems.gram

ControlSystems.grampd

ControlSystems.hinfnorm

ControlSystems.innovation_form

ControlSystems.linfnorm

ControlSystems.margin

ControlSystems.markovparam

ControlSystems.observer_controller

ControlSystems.observer_predictor

ControlSystems.obsv

ControlSystems.poles

ControlSystems.reduce_sys

ControlSystems.relative_gain_array

ControlSystems.relative_gain_array

ControlSystems.similarity_transform

ControlSystems.tzeros

ControlSystems.zpkdata

LinearAlgebra.lyap

LinearAlgebra.norm

Analysis

ControlSystems.damp — Method

Wn, zeta, ps = damp(sys)

Compute the natural frequencies, Wn, and damping ratios, zeta, of the poles, ps, of sys

ControlSystems.dampreport — Method

dampreport(sys)

Display a report of the poles, damping ratio, natural frequency, and time constant of the system sys

ControlSystems.dcgain — Function

dcgain(sys, ϵ=0)

Compute the dcgain of system sys.

equal to G(0) for continuous-time systems and G(1) for discrete-time systems.

ϵ can be provided to evaluate the dcgain with a small perturbation into the stability region of the complex plane.

ControlSystems.delaymargin — Method

dₘ = delaymargin(G::LTISystem)

Only supports SISO systems

ControlSystems.gangoffour — Method

S, PS, CS, T = gangoffour(P, C; minimal=true)
gangoffour(P::AbstractVector, C::AbstractVector; minimal=true)

Given a transfer function describing the plant P and a transfer function describing the controller C, computes the four transfer functions in the Gang-of-Four.

S = 1/(1+PC) Sensitivity function

PS = P/(1+PC) Load disturbance to measurement signal

CS = C/(1+PC) Measurement noise to control signal

T = PC/(1+PC) Complementary sensitivity function

Only supports SISO systems

ControlSystems.gangofseven — Method

S, PS, CS, T, RY, RU, RE = gangofseven(P,C,F)

Given transfer functions describing the Plant P, the controller C and a feed forward block F, computes the four transfer functions in the Gang-of-Four and the transferfunctions corresponding to the feed forward.

S = 1/(1+PC) Sensitivity function

PS = P/(1+PC)

CS = C/(1+PC)

T = PC/(1+PC) Complementary sensitivity function

RY = PCF/(1+PC)

RU = CF/(1+P*C)

RE = F/(1+P*C)

Only supports SISO systems

ControlSystems.margin — Method

ωgₘ, gₘ, ωϕₘ, ϕₘ = margin(sys::LTISystem, w::Vector; full=false, allMargins=false)

returns frequencies for gain margins, gain margins, frequencies for phase margins, phase margins

If !allMargins, return only the smallest margin

If full return also fullPhase

ControlSystems.markovparam — Method

markovparam(sys, n)

Compute the nth markov parameter of state-space system sys. This is defined as the following:

h(0) = D

h(n) = C*A^(n-1)*B

ControlSystems.poles — Method

poles(sys)

Compute the poles of system sys.

ControlSystems.reduce_sys — Method

Implements REDUCE in the Emami-Naeini & Van Dooren paper. Returns transformed A, B, C, D matrices. These are empty if there are no zeros.

ControlSystems.relative_gain_array — Method

relative_gain_array(A::AbstractMatrix; tol = 1.0e-15)

Reference: "On the Relative Gain Array (RGA) with Singular and Rectangular Matrices" Jeffrey Uhlmann https://arxiv.org/pdf/1805.10312.pdf

ControlSystems.relative_gain_array — Method

relative_gain_array(G, w::AbstractVector)
relative_gain_array(G, w::Number)

Calculate the relative gain array of G at frequencies w. G(iω) .* pinv(tranpose(G(iω)))

The RGA can be used to find input-output pairings for MIMO control using individially tuned loops. Pair the inputs and outputs such that the RGA(ωc) at the crossover frequency becomes as close to diagonal as possible. Avoid pairings such that RGA(0) contains negative diagonal elements.

The sum of the absolute values of the entries in the RGA is a good measure of the "true condition number" of G, the best condition number that can be achieved by input/output scaling of G, -Glad, Ljung.

The RGA is invariant to input/output scaling of G.

If the RGA contains large entries, the system may be sensitive to model errors, -Skogestad, "Multivariable Feedback Control: Analysis and Design":

Uncertainty in the input channels (diagonal input uncertainty). Plants with

large RGA-elements around the crossover frequency are fundamentally difficult to control because of sensitivity to input uncertainty (e.g. caused by uncertain or neglected actuator dynamics). In particular, decouplers or other inverse-based controllers should not be used for plants with large RGAeleme

Element uncertainty. Large RGA-elements imply sensitivity to element-by-element uncertainty.

However, this kind of uncertainty may not occur in practice due to physical couplings between the transfer function elements. Therefore, diagonal input uncertainty (which is always present) is usually of more concern for plants with large RGAelemen

The relative gain array is computed using the The unit-consistent (UC) generalized inverse Reference: "On the Relative Gain Array (RGA) with Singular and Rectangular Matrices" Jeffrey Uhlmann https://arxiv.org/pdf/1805.10312.pdf

ControlSystems.tzeros — Method

tzeros(sys)

Compute the invariant zeros of the system sys. If sys is a minimal realization, these are also the transmission zeros.

ControlSystems.zpkdata — Method

z, p, k = zpkdata(sys)

Compute the zeros, poles, and gains of system sys.

Returns

z : Matrix{Vector{ComplexF64}}, (ny x nu)

p : Matrix{Vector{ComplexF64}}, (ny x nu)

k : Matrix{Float64}, (ny x nu)

ControlSystems.are — Method

are(::Continuous, A, B, Q, R)

Compute 'X', the solution to the continuous-time algebraic Riccati equation, defined as A'X + XA - (XB)R^-1(B'X) + Q = 0, where R is non-singular.

Uses MatrixEquations.arec.

ControlSystems.are — Method

are(::Discrete, A, B, Q, R; kwargs...)

Compute X, the solution to the discrete-time algebraic Riccati equation, defined as A'XA - X - (A'XB)(B'XB + R)^-1(B'XA) + Q = 0, where Q>=0 and R>0

Uses MatrixEquations.ared. For keyword arguments, see the docstring of ControlSystems.MatrixEquations.ared

ControlSystems.balance — Function

S, P, B = balance(A[, perm=true])

Compute a similarity transform T = S*P resulting in B = T\A*T such that the row and column norms of B are approximately equivalent. If perm=false, the transformation will only scale A using diagonal S, and not permute A (i.e., set P=I).

ControlSystems.balreal — Method

sysr, G, T = balreal(sys::StateSpace)

Calculates a balanced realization of the system sys, such that the observability and reachability gramians of the balanced system are equal and diagonal G. T is the similarity transform between the old state x and the new state z such that Tz = x.

See also gram, baltrunc

Reference: Varga A., Balancing-free square-root algorithm for computing singular perturbation approximations.

ControlSystems.baltrunc — Method

sysr, G, T = baltrunc(sys::StateSpace; atol = √ϵ, rtol=1e-3, n = nothing, residual = false)

Reduces the state dimension by calculating a balanced realization of the system sys, such that the observability and reachability gramians of the balanced system are equal and diagonal G, and truncating it to order n. If n is not provided, it's chosen such that all states corresponding to singular values less than atol and less that rtol σmax are removed.

T is the similarity transform between the old state x and the newstate z such that Tz = x.

If residual = true, matched static gain is achieved through "residualization", i.e., setting

\[0 = A_{21}x_{1} + A_{22}x_{2} + B_{2}u\]

where indices 1/2 correspond to the remaining/truncated states respectively.

See also gram, balreal

Glad, Ljung, Reglerteori: Flervariabla och Olinjära metoder

ControlSystems.covar — Method

P = covar(sys, W)

Calculate the stationary covariance P = E[y(t)y(t)'] of the output y of a StateSpace model sys driven by white Gaussian noise w with covariance E[w(t)w(τ)]=W*δ(t-τ) (δ is the Dirac delta).

Remark: If sys is unstable then the resulting covariance is a matrix of Infs. Entries corresponding to direct feedthrough (DWD' .!= 0) will equal Inf for continuous-time systems.

ControlSystems.ctrb — Method

ctrb(A, B) or ctrb(sys)

Compute the controllability matrix for the system described by (A, B) or sys.

Note that checking for controllability by computing the rank from ctrb is not the most numerically accurate way, a better method is checking if gram(sys, :c) is positive definite.

ControlSystems.gram — Method

gram(sys, opt; kwargs...)

Compute the grammian of system sys. If opt is :c, computes the controllability grammian. If opt is :o, computes the observability grammian.

See also grampd For keyword arguments, see grampd.

ControlSystems.grampd — Method

U = grampd(sys, opt; kwargs...)

Return a Cholesky factor U of the grammian of system sys. If opt is :c, computes the controllability grammian G = U*U'. If opt is :o, computes the observability grammian G = U'U.

Obtain a Cholesky object by Cholesky(U) for observability grammian

Uses MatrixEquations.plyapc/plyapd. For keyword arguments, see the docstring of ControlSystems.MatrixEquations.plyapc/plyapd

ControlSystems.hinfnorm — Method

Ninf, ω_peak = hinfnorm(sys; tol=1e-6)

Compute the H∞ norm Ninf of the LTI system sys, together with a frequency ω_peak at which the gain Ninf is achieved.

Ninf := sup_ω σ_max[sys(iω)] if G is stable (σ_max = largest singular value) := Inf' ifG` is unstable

tol is an optional keyword argument for the desired relative accuracy for the computed H∞ norm (not an absolute certificate).

sys is first converted to a state space model if needed.

The continuous-time L∞ norm computation implements the 'two-step algorithm' in:
N.A. Bruinsma and M. Steinbuch, 'A fast algorithm to compute the H∞-norm of a transfer function matrix', Systems and Control Letters (1990), pp. 287-293.

For the discrete-time version, see:
P. Bongers, O. Bosgra, M. Steinbuch, 'L∞-norm calculation for generalized state space systems in continuous and discrete time', American Control Conference, 1991.

See also linfnorm.

ControlSystems.innovation_form — Method

sysi = innovation_form(sys, R1, R2)
sysi = innovation_form(sys; sysw=I, syse=I, R1=I, R2=I)

Takes a system

x' = Ax + Bu + w ~ R1
y  = Cx + e ~ R2

and returns the system

x' = Ax + Kv
y  = Cx + v

where v is the innovation sequence.

If sysw (syse) is given, the covariance resulting in filtering noise with R1 (R2) through sysw (syse) is used as covariance.

See Stochastic Control, Chapter 4, Åström

ControlSystems.linfnorm — Method

Ninf, ω_peak = linfnorm(sys; tol=1e-6)

Compute the L∞ norm Ninf of the LTI system sys, together with a frequency ω_peak at which the gain Ninf is achieved.

Ninf := sup_ω σ_max[sys(iω)] (σ_max denotes the largest singular value)

tol is an optional keyword argument representing the desired relative accuracy for the computed L∞ norm (this is not an absolute certificate however).

sys is first converted to a state space model if needed.

The continuous-time L∞ norm computation implements the 'two-step algorithm' in:
N.A. Bruinsma and M. Steinbuch, 'A fast algorithm to compute the H∞-norm of a transfer function matrix', Systems and Control Letters (1990), pp. 287-293.

For the discrete-time version, see:
P. Bongers, O. Bosgra, M. Steinbuch, 'L∞-norm calculation for generalized state space systems in continuous and discrete time', American Control Conference, 1991.

See also hinfnorm.

ControlSystems.observer_controller — Method

cont = observer_controller(sys, L::AbstractMatrix, K::AbstractMatrix)

Return the observer_controller cont that is given by ss(A - B*L - K*C + K*D*L, K, L, 0)

Such that feedback(sys, cont) produces a closed-loop system with eigenvalues given by A-KC and A-BL.

Arguments:

sys: Model of system

L: State-feedback gain u = -Lx

K: Observer gain

ControlSystems.observer_predictor — Method

observer_predictor(sys::AbstractStateSpace, R1, R2)
observer_predictor(sys::AbstractStateSpace, K)

Return the observer_predictor system x̂' = (A - KC)x̂ + (B-KD)u + Ky ŷ = Cx + Du with the input equation [B-KD K] * [u; y]

If covariance matrices R1, R2 are given, the kalman gain K is calculaded.

See also innovation_form.

ControlSystems.obsv — Function

obsv(A, C, n=size(A,1))
obsv(sys, n=sys.nx)

Compute the observability matrix with n rows for the system described by (A, C) or sys. Providing the optional n > sys.nx returns an extended observability matrix.

Note that checking for observability by computing the rank from obsv is not the most numerically accurate way, a better method is checking if gram(sys, :o) is positive definite.

ControlSystems.similarity_transform — Method

syst = similarity_transform(sys, T; unitary=false)

Perform a similarity transform T : Tx̃ = x on sys such that

Ã = T⁻¹AT
B̃ = T⁻¹ B
C̃ = CT
D̃ = D

If unitary=true, T is assumed unitary and the matrix adjoint is used instead of the inverse.

LinearAlgebra.lyap — Method

lyap(A, Q; kwargs...)

Compute the solution X to the discrete Lyapunov equation AXA' - X + Q = 0.

Uses MatrixEquations.lyapc / MatrixEquations.lyapd. For keyword arguments, see the docstring of ControlSystems.MatrixEquations.lyapc / ControlSystems.MatrixEquations.lyapd

LinearAlgebra.norm — Function

norm(sys, p=2; tol=1e-6)

norm(sys) or norm(sys,2) computes the H2 norm of the LTI system sys.

norm(sys, Inf) computes the H∞ norm of the LTI system sys. The H∞ norm is the same as the H∞ for stable systems, and Inf for unstable systems. If the peak gain frequency is required as well, use the function hinfnorm instead. See hinfnorm for further documentation.

tol is an optional keyword argument, used only for the computation of L∞ norms. It represents the desired relative accuracy for the computed L∞ norm (this is not an absolute certificate however).

sys is first converted to a StateSpace model if needed.