ControlSystems.balance

ControlSystems.balreal

ControlSystems.baltrunc

ControlSystems.care

ControlSystems.covar

ControlSystems.ctrb

ControlSystems.damp

ControlSystems.dampreport

ControlSystems.dare

ControlSystems.dcgain

ControlSystems.delaymargin

ControlSystems.dlyap

ControlSystems.gangoffour

ControlSystems.gangofseven

ControlSystems.gram

ControlSystems.hinfnorm

ControlSystems.innovation_form

ControlSystems.linfnorm

ControlSystems.margin

ControlSystems.markovparam

ControlSystems.obsv

ControlSystems.pole

ControlSystems.reduce_sys

ControlSystems.similarity_transform

ControlSystems.tzero

ControlSystems.zpkdata

LinearAlgebra.norm

Analysis

ControlSystems.damp — Method

Wn, zeta, ps = damp(sys)

Compute the natural frequencies, Wn, and damping ratios, zeta, of the poles, ps, of sys

ControlSystems.dampreport — Method

dampreport(sys)

Display a report of the poles, damping ratio, natural frequency, and time constant of the system sys

ControlSystems.dcgain — Method

dcgain(sys)

Compute the dcgain of system sys.

equal to G(0) for continuous-time systems and G(1) for discrete-time systems.

ControlSystems.delaymargin — Method

dₘ = delaymargin(G::LTISystem)

Only supports SISO systems

ControlSystems.gangoffour — Method

S,D,N,T = gangoffour(P,C; minimal=true)
gangoffour(P::AbstractVector,C::AbstractVector; minimal=true)

Given a transfer function describing the Plant P and a transfer function describing the controller C, computes the four transfer functions in the Gang-of-Four.

minimal determines whether or not to call minreal on the computed systems.

S = 1/(1+PC) Sensitivity function

D = P/(1+PC)

N = C/(1+PC)

T = PC/(1+PC) Complementary sensitivity function

Only supports SISO systems

ControlSystems.gangofseven — Method

S, D, N, T, RY, RU, RE = gangofseven(P,C,F)

Given transfer functions describing the Plant P, the controller C and a feed forward block F, computes the four transfer functions in the Gang-of-Four and the transferfunctions corresponding to the feed forward.

S = 1/(1+PC) Sensitivity function

D = P/(1+PC)

N = C/(1+PC)

T = PC/(1+PC) Complementary sensitivity function

RY = PCF/(1+PC)

RU = CF/(1+P*C)

RE = F/(1+P*C)

Only supports SISO systems

ControlSystems.margin — Method

ωgₘ, gₘ, ωϕₘ, ϕₘ = margin{S<:Real}(sys::LTISystem, w::AbstractVector{S}; full=false, allMargins=false)

returns frequencies for gain margins, gain margins, frequencies for phase margins, phase margins

If !allMargins, return only the smallest margin

If full return also fullPhase

ControlSystems.markovparam — Method

markovparam(sys, n)

Compute the nth markov parameter of state-space system sys. This is defined as the following:

h(0) = D

h(n) = C*A^(n-1)*B

ControlSystems.pole — Method

pole(sys)

Compute the poles of system sys.

ControlSystems.reduce_sys — Method

Implements REDUCE in the Emami-Naeini & Van Dooren paper. Returns transformed A, B, C, D matrices. These are empty if there are no zeros.

ControlSystems.tzero — Method

tzero(sys)

Compute the invariant zeros of the system sys. If sys is a minimal realization, these are also the transmission zeros.

ControlSystems.zpkdata — Method

z, p, k = zpkdata(sys)

Compute the zeros, poles, and gains of system sys.

Returns

z : Matrix{Vector{ComplexF64}}, (ny x nu)

p : Matrix{Vector{ComplexF64}}, (ny x nu)

k : Matrix{Float64}, (ny x nu)

ControlSystems.balance — Function

S, P, B = balance(A[, perm=true])

Compute a similarity transform T = S*P resulting in B = T\A*T such that the row and column norms of B are approximately equivalent. If perm=false, the transformation will only scale A using diagonal S, and not permute A (i.e., set P=I).

ControlSystems.balreal — Method

sysr, G = balreal(sys::StateSpace)

Calculates a balanced realization of the system sys, such that the observability and reachability gramians of the balanced system are equal and diagonal G

See also gram, baltrunc

Glad, Ljung, Reglerteori: Flervariabla och Olinjära metoder

ControlSystems.baltrunc — Method

sysr, G = baltrunc(sys::StateSpace; atol = √ϵ, rtol=1e-3, unitgain=true, n = nothing)

Reduces the state dimension by calculating a balanced realization of the system sys, such that the observability and reachability gramians of the balanced system are equal and diagonal G, and truncating it to order n. If n is not provided, it's chosen such that all states corresponding to singular values less than atol and less that rtol σmax are removed.

If unitgain=true, the matrix D is chosen such that unit static gain is achieved.

See also gram, balreal

Glad, Ljung, Reglerteori: Flervariabla och Olinjära metoder

ControlSystems.care — Method

care(A, B, Q, R)

Compute 'X', the solution to the continuous-time algebraic Riccati equation, defined as A'X + XA - (XB)R^-1(B'X) + Q = 0, where R is non-singular.

Algorithm taken from: Laub, "A Schur Method for Solving Algebraic Riccati Equations." http://dspace.mit.edu/bitstream/handle/1721.1/1301/R-0859-05666488.pdf

ControlSystems.covar — Method

P = covar(sys, W)

Calculate the stationary covariance P = E[y(t)y(t)'] of the output y of a StateSpace model sys driven by white Gaussian noise w with covariance E[w(t)w(τ)]=W*δ(t-τ) (δ is the Dirac delta).

Remark: If sys is unstable then the resulting covariance is a matrix of Infs. Entries corresponding to direct feedthrough (DWD' .!= 0) will equal Inf for continuous-time systems.

ControlSystems.ctrb — Method

ctrb(A, B) or ctrb(sys)

Compute the controllability matrix for the system described by (A, B) or sys.

Note that checking for controllability by computing the rank from ctrb is not the most numerically accurate way, a better method is checking if gram(sys, :c) is positive definite.

ControlSystems.dare — Method

dare(A, B, Q, R)

Compute X, the solution to the discrete-time algebraic Riccati equation, defined as A'XA - X - (A'XB)(B'XB + R)^-1(B'XA) + Q = 0, where Q>=0 and R>0

Algorithm taken from: Laub, "A Schur Method for Solving Algebraic Riccati Equations." http://dspace.mit.edu/bitstream/handle/1721.1/1301/R-0859-05666488.pdf

ControlSystems.dlyap — Method

dlyap(A, Q)

Compute the solution X to the discrete Lyapunov equation AXA' - X + Q = 0.

ControlSystems.gram — Method

gram(sys, opt)

Compute the grammian of system sys. If opt is :c, computes the controllability grammian. If opt is :o, computes the observability grammian.

ControlSystems.hinfnorm — Method

(Ninf, ω_peak) = hinfnorm(sys; tol=1e-6)

Compute the H∞ norm Ninf of the LTI system sys, together with a frequency ω_peak at which the gain Ninf is achieved.

Ninf := sup_ω σ_max[sys(iω)] if G is stable (σ_max = largest singular value) := Inf' ifG` is unstable

tol is an optional keyword argument for the desired relative accuracy for the computed H∞ norm (not an absolute certificate).

sys is first converted to a state space model if needed.

The continuous-time L∞ norm computation implements the 'two-step algorithm' in:
N.A. Bruinsma and M. Steinbuch, 'A fast algorithm to compute the H∞-norm of a transfer function matrix', Systems and Control Letters (1990), pp. 287-293.

For the discrete-time version, see:
P. Bongers, O. Bosgra, M. Steinbuch, 'L∞-norm calculation for generalized state space systems in continuous and discrete time', American Control Conference, 1991.

See also linfnorm.

ControlSystems.innovation_form — Method

sysi = innovationform(sys, R1, R2) sysi = innovationform(sys; sysw=I, syse=I, R1=I, R2=I)

Takes a system

x' = Ax + Bu + w ~ R1
y  = Cx + e ~ R2

and returns the system

x' = Ax + Kv
y  = Cx + v

where v is the innovation sequence.

If sysw (syse) is given, the covariance resulting in filtering noise with R1 (R2) through sysw (syse) is used as covariance.

See Stochastic Control, Chapter 4, Åström

ControlSystems.linfnorm — Method

(Ninf, ω_peak) = linfnorm(sys; tol=1e-6)

Compute the L∞ norm Ninf of the LTI system sys, together with a frequency ω_peak at which the gain Ninf is achieved.

Ninf := sup_ω σ_max[sys(iω)] (σ_max denotes the largest singular value)

tol is an optional keyword argument representing the desired relative accuracy for the computed L∞ norm (this is not an absolute certificate however).

sys is first converted to a state space model if needed.

The continuous-time L∞ norm computation implements the 'two-step algorithm' in:
N.A. Bruinsma and M. Steinbuch, 'A fast algorithm to compute the H∞-norm of a transfer function matrix', Systems and Control Letters (1990), pp. 287-293.

For the discrete-time version, see:
P. Bongers, O. Bosgra, M. Steinbuch, 'L∞-norm calculation for generalized state space systems in continuous and discrete time', American Control Conference, 1991.

See also hinfnorm.

ControlSystems.obsv — Method

obsv(A, C) or obsv(sys)

Compute the observability matrix for the system described by (A, C) or sys.

Note that checking for observability by computing the rank from obsv is not the most numerically accurate way, a better method is checking if gram(sys, :o) is positive definite.

ControlSystems.similarity_transform — Method

syst = similarity_transform(sys, T)

Perform a similarity transform T : Tx̃ = x on sys such that

Ã = T⁻¹AT
B̃ = T⁻¹ B
C̃ = CT
D̃ = D

LinearAlgebra.norm — Function

.. norm(sys, p=2; tol=1e-6)

norm(sys) or norm(sys,2) computes the H2 norm of the LTI system sys.

norm(sys, Inf) computes the H∞ norm of the LTI system sys. The H∞ norm is the same as the H∞ for stable systems, and Inf for unstable systems. If the peak gain frequency is required as well, use the function hinfnorm instead. See hinfnorm for further documentation.

tol is an optional keyword argument, used only for the computation of L∞ norms. It represents the desired relative accuracy for the computed L∞ norm (this is not an absolute certificate however).

sys is first converted to a StateSpace model if needed.