ControlSystems.append

ControlSystems.bodev

ControlSystems.bodev

ControlSystems.c2d

ControlSystems.c2d_poly2poly

ControlSystems.c2d_roots2poly

ControlSystems.dab

ControlSystems.dkalman

ControlSystems.dlqr

ControlSystems.feedback

ControlSystems.feedback

ControlSystems.feedback

ControlSystems.feedback2dof

ControlSystems.kalman

ControlSystems.laglink

ControlSystems.leadlink

ControlSystems.leadlinkat

ControlSystems.leadlinkcurve

ControlSystems.lft

ControlSystems.loopshapingPI

ControlSystems.lqr

ControlSystems.nyquistv

ControlSystems.nyquistv

ControlSystems.parallel

ControlSystems.pid

ControlSystems.pidplots

ControlSystems.place

ControlSystems.rlocus

ControlSystems.rstc

ControlSystems.rstd

ControlSystems.series

ControlSystems.sigmav

ControlSystems.sigmav

ControlSystems.sminreal

ControlSystems.stabregionPID

ControlSystems.starprod

ControlSystems.zpconv

Synthesis

ControlSystems.dkalman — Method

dkalman(A, C, R1, R2) kalman(sys, R1, R2)`

Calculate the optimal Kalman gain for discrete time systems

ControlSystems.dlqr — Method

dlqr(A, B, Q, R), dlqr(sys, Q, R)

Calculate the optimal gain matrix K for the state-feedback law u[k] = K*x[k] that minimizes the cost function:

J = sum(x'Qx + u'Ru, 0, inf).

For the discrte time model x[k+1] = Ax[k] + Bu[k].

See also lqg

Usage example:

using LinearAlgebra # For identity matrix I
h = 0.1
A = [1 h; 0 1]
B = [0;1]
C = [1 0]
sys = ss(A,B,C,0, h)
Q = I
R = I
L = dlqr(A,B,Q,R) # lqr(sys,Q,R) can also be used

u(x,t) = -L*x # Form control law,
t=0:h:5
x0 = [1,0]
y, t, x, uout = lsim(sys,u,t,x0=x0)
plot(t,x, lab=["Position"  "Velocity"], xlabel="Time [s]")

ControlSystems.kalman — Method

kalman(A, C, R1, R2) kalman(sys, R1, R2)`

Calculate the optimal Kalman gain

See also LQG

ControlSystems.lqr — Method

lqr(A, B, Q, R)

Calculate the optimal gain matrix K for the state-feedback law u = K*x that minimizes the cost function:

J = integral(x'Qx + u'Ru, 0, inf).

For the continuous time model dx = Ax + Bu.

lqr(sys, Q, R)

Solve the LQR problem for state-space system sys. Works for both discrete and continuous time systems.

See also LQG

Usage example:

using LinearAlgebra # For identity matrix I
A = [0 1; 0 0]
B = [0;1]
C = [1 0]
sys = ss(A,B,C,0)
Q = I
R = I
L = lqr(sys,Q,R)

u(x,t) = -L*x # Form control law,
t=0:0.1:5
x0 = [1,0]
y, t, x, uout = lsim(sys,u,t,x0=x0)
plot(t,x, lab=["Position" "Velocity"], xlabel="Time [s]")

ControlSystems.place — Method

place(A, B, p), place(sys::StateSpace, p)

Calculate gain matrix K such that the poles of (A-BK) in are in p.

Uses Ackermann's formula.

ControlSystems.c2d — Function

[sysd, x0map] = c2d(sys, Ts, method=:zoh)

Convert the continuous system sys into a discrete system with sample time Ts, using the provided method. Currently only :zoh and :foh are provided.

Returns the discrete system sysd, and a matrix x0map that transforms the initial conditions to the discrete domain by x0_discrete = x0map*[x0; u0]

ControlSystems.c2d_poly2poly — Method

c2d_poly2poly(ro,h)

returns the polynomial coefficients in discrete time given polynomial coefficients in continuous time

ControlSystems.c2d_roots2poly — Method

c2d_roots2poly(ro,h)

returns the polynomial coefficients in discrete time given a vector of roots in continuous time

ControlSystems.dab — Method

DAB Solves the Diophantine-Aryabhatta-Bezout identity

X,Y = DAB(A,B,C)

AX + BY = C, where A, B, C, X and Y are polynomials and deg Y = deg A - 1.

See Computer-Controlled Systems: Theory and Design, Third Edition Karl Johan Åström, Björn Wittenmark

ControlSystems.rstc — Method

See ?rstd for the discerte case

ControlSystems.rstd — Method

rstd Polynomial synthesis in discrete time.

R,S,T=rstd(BPLUS,BMINUS,A,BM1,AM,AO,AR,AS)

R,S,T=rstd(BPLUS,BMINUS,A,BM1,AM,AO,AR)

R,S,T=rstd(BPLUS,BMINUS,A,BM1,AM,AO)

Polynomial synthesis according to CCS ch 10 to design a controller R(q) u(k) = T(q) r(k) - S(q) y(k)

Inputs: BPLUS : Part of open loop numerator BMINUS : Part of open loop numerator A : Open loop denominator BM1 : Additional zeros AM : Closed loop denominator AO : Observer polynomial AR : Pre-specified factor of R, e.g integral part [1, -1]^k AS : Pre-specified factor of S, e.g notch filter [1, 0, w^2]

Outputs: R,S,T : Polynomials in controller

See function DAB how the solution to the Diophantine- Aryabhatta-Bezout identity is chosen.

See Computer-Controlled Systems: Theory and Design, Third Edition Karl Johan Åström, Björn Wittenmark

ControlSystems.zpconv — Method

zpc(a,r,b,s) form conv(a,r) + conv(b,s) where the lengths of the polynomials are equalized by zero-padding such that the addition can be carried out

ControlSystems.laglink — Method

laglink(a, M; h=0)

Returns a phase retarding link, the rule of thumb a = 0.1ωc guarantees less than 6 degrees phase margin loss. The bode curve will go from M, bend down at a/M and level out at 1 for frequencies > a

ControlSystems.leadlink — Method

leadlink(b, N, K; h=0)

Returns a phase advancing link, the top of the phase curve is located at ω = b√(N) where the link amplification is K√(N) The bode curve will go from K, bend up at b and level out at KN for frequencies > bN

The phase advance at ω = b√(N) can be plotted as a function of N with leadlinkcurve()

Values of N < 1 will give a phase retarding link.

See also leadlinkat laglink

ControlSystems.leadlinkat — Method

leadlinkat(ω, N, K; h=0)

Returns a phase advancing link, the top of the phase curve is located at ω where the link amplification is K√(N) The bode curve will go from K, bend up at ω/√(N) and level out at KN for frequencies > ω√(N)

The phase advance at ω can be plotted as a function of N with leadlinkcurve()

Values of N < 1 will give a phase retarding link.

See also leadlink laglink

ControlSystems.leadlinkcurve — Function

Plot the phase advance as a function of N for a lead link (phase advance link)

If an input argument s is given, the curve is plotted from s to 10, else from 1 to 10.

See also Leadlink, leadlinkat

ControlSystems.loopshapingPI — Method

kp,ki,C = loopshapingPI(P,ωp; ϕl,rl, phasemargin, doplot = false)

Selects the parameters of a PI-controller such that the Nyquist curve of P at the frequency ωp is moved to rl exp(i ϕl)

If phasemargin is supplied, ϕl is selected such that the curve is moved to an angle of phasemargin - 180 degrees

If no rl is given, the magnitude of the curve at ωp is kept the same and only the phase is affected, the same goes for ϕl if no phasemargin is given.

Set doplot = true to plot the gangoffourplot and nyquistplot of the system.

See also pidplots, stabregionPID

ControlSystems.pid — Method

Calculates and returns a PID controller on transfer function form. time indicates whether or not the parameters are given as gains (default) or as time constants series indicates whether or not the series form or parallel form (default) is desired

C = pid(; kp=0, ki=0; kd=0, time=false, series=false)

ControlSystems.pidplots — Method

Plots interesting figures related to closing the loop around process P with a PID controller Send in a bunch of PID-parameters in any of the vectors kp, ki, kd. The vectors must be the same length.

time indicates whether or not the parameters are given as gains (default) or as time constants

series indicates whether or not the series form or parallel form (default) is desired

Available plots are :gof for Gang of four, :nyquist, :controller for a bode plot of the controller TF and :pz for pole-zero maps

One can also supply a frequency vector ω to be used in Bode and Nyquist plots

pidplots(P, args...; kps=0, kis=0, kds=0, time=false, series=false, ω=0)

See also loopshapingPI, stabregionPID

ControlSystems.rlocus — Function

rlocusplot(P::LTISystem, K)

Computes and plots the root locus of the SISO LTISystem P with a negative feedback loop and feedback gains K, if K is not provided, range(1e-6,stop=500,length=10000) is used. If OrdinaryDiffEq.jl is installed and loaded by the user (using OrdinaryDiffEq), rlocusplot will use an adaptive step-size algorithm to select values of K. A scalar Kmax can then be given as second argument.

ControlSystems.stabregionPID — Function

fig, kp, ki = stabregionPID(P, [ω]; kd=0, doplot = true)

Segments of the curve generated by this program is the boundary of the stability region for a process with transfer function P(s) The PID controller is assumed to be on the form kp +ki/s +kd s

The curve is found by analyzing P(s)*C(s) = -1 ⟹ |PC| = |P| |C| = 1 arg(P) + arg(C) = -π

If P is a function (e.g. s -> exp(-sqrt(s)) ), the stability of feedback loops using PI-controllers can be analyzed for processes with models with arbitrary analytic functions

See also stabregionPID, loopshapingPI, pidplots

ControlSystems.sminreal — Method

sminreal(sys)

Compute the structurally minimal realization of the state-space system sys. A structurally minimal realization is one where only states that can be determined to be uncontrollable and unobservable based on the location of 0s in sys are removed.

ControlSystems.append — Method

append(systems::StateSpace...), append(systems::TransferFunction...)

Append systems in block diagonal form

ControlSystems.feedback — Method

feedback(s1::AbstractStateSpace, s2::AbstractStateSpace;
         U1=:, Y1=:, U2=:, Y2=:, W1=:, Z1=:, W2=Int[], Z2=Int[],
         Wperm=:, Zperm=:, pos_feedback::Bool=false)

U1, Y1, U2, Y2 contain the indices of the signals that should be connected. W1, Z1, W2, Z2 contain the signal indices of s1 and s2 that should be kept.

Specify Wperm and Zperm to reorder [w1; w2] and [z1; z2] in the resulting statespace model.

Negative feedback is the default. Specify pos_feedback=true for positive feedback.

See Zhou etc. for similar (somewhat less symmetric) formulas.

ControlSystems.feedback — Method

feedback(L) Returns L/(1+L) feedback(P1,P2) Returns P1/(1+P1*P2)

ControlSystems.feedback — Method

feedback(sys)

feedback(sys1,sys2)

Forms the negative feedback interconnection

>-+ sys1 +-->
  |      |
 (-)sys2 +

If no second system is given, negative identity feedback is assumed

ControlSystems.feedback2dof — Method

feedback2dof(P,R,S,T) Return BT/(AR+ST) where B and A are the numerator and denomenator polynomials of P respectively feedback2dof(B,A,R,S,T) Return BT/(AR+ST)

ControlSystems.lft — Function

lft(G, Δ, type=:l)

Lower and upper linear fractional transformation between systems G and Δ.

Specify :l lor lower LFT, and :u for upper LFT.

G must have more inputs and outputs than Δ has outputs and inputs.

For details, see Chapter 9.1 in Zhou, K. and JC Doyle. Essentials of robust control, Prentice hall (NJ), 1998

ControlSystems.parallel — Method

parallel(sys1::LTISystem, sys2::LTISystem)

Connect systems in parallel, equivalent to sys2+sys1

ControlSystems.series — Method

series(sys1::LTISystem, sys2::LTISystem)

Connect systems in series, equivalent to sys2*sys1

ControlSystems.starprod — Method

starprod(sys1, sys2, dimu, dimy)

Compute the Redheffer star product.

length(U1) = length(Y2) = dimu and length(Y1) = length(U2) = dimy

For details, see Chapter 9.3 in Zhou, K. and JC Doyle. Essentials of robust control, Prentice hall (NJ), 1998

ControlSystems.bodev — Method

ControlSystems.bodev(sys::LTISystem, w::AbstractVector; )

For use with SISO systems where it acts the same as ControlSystems.bode but with the extra dimensions removed in the returned values.

ControlSystems.bodev — Method

ControlSystems.bodev(sys::LTISystem; )

For use with SISO systems where it acts the same as ControlSystems.bode but with the extra dimensions removed in the returned values.

ControlSystems.nyquistv — Method

ControlSystems.nyquistv(sys::LTISystem, w::AbstractVector; )

For use with SISO systems where it acts the same as ControlSystems.nyquist but with the extra dimensions removed in the returned values.

ControlSystems.nyquistv — Method

ControlSystems.nyquistv(sys::LTISystem; )

For use with SISO systems where it acts the same as ControlSystems.nyquist but with the extra dimensions removed in the returned values.

ControlSystems.sigmav — Method

ControlSystems.sigmav(sys::LTISystem, w::AbstractVector; )

For use with SISO systems where it acts the same as ControlSystems.sigma but with the extra dimensions removed in the returned values.

ControlSystems.sigmav — Method

ControlSystems.sigmav(sys::LTISystem; )

For use with SISO systems where it acts the same as ControlSystems.sigma but with the extra dimensions removed in the returned values.