ControlSystems.G_CS

ControlSystems.G_PS

ControlSystems.add_input

ControlSystems.add_output

ControlSystems.append

ControlSystems.array2mimo

ControlSystems.bodev

ControlSystems.bodev

ControlSystems.c2d

ControlSystems.c2d_poly2poly

ControlSystems.c2d_roots2poly

ControlSystems.c2d_x0map

ControlSystems.comp_sensitivity

ControlSystems.d2c

ControlSystems.dab

ControlSystems.feedback

ControlSystems.feedback

ControlSystems.feedback2dof

ControlSystems.feedback2dof

ControlSystems.freqrespv

ControlSystems.freqrespv

ControlSystems.freqrespv

ControlSystems.input_comp_sensitivity

ControlSystems.input_sensitivity

ControlSystems.kalman

ControlSystems.laglink

ControlSystems.leadlink

ControlSystems.leadlinkat

ControlSystems.leadlinkcurve

ControlSystems.lft

ControlSystems.loopshapingPI

ControlSystems.lqr

ControlSystems.nyquistv

ControlSystems.nyquistv

ControlSystems.output_comp_sensitivity

ControlSystems.output_sensitivity

ControlSystems.parallel

ControlSystems.pid

ControlSystems.pidplots

ControlSystems.place

ControlSystems.placePI

ControlSystems.rlocusplot

ControlSystems.rstc

ControlSystems.rstd

ControlSystems.sensitivity

ControlSystems.series

ControlSystems.sigmav

ControlSystems.sigmav

ControlSystems.sminreal

ControlSystems.stabregionPID

ControlSystems.starprod

ControlSystems.zpconv

Synthesis

For $H_\infty$ and $H_2$ synthesis as well as more advanced LQG design, see RobustAndOptimalControl.

ControlSystems.kalman — Method

kalman(Continuous, A, C, R1, R2)
kalman(Discrete, A, C, R1, R2)
kalman(sys, R1, R2)

Calculate the optimal Kalman gain

The args...; kwargs... are sent to the Riccati solver, allowing specification of cross-covariance etc. See ?MatrixEquations.arec/ared for more help.

ControlSystems.lqr — Method

lqr(sys, Q, R)
lqr(Continuous, A, B, Q, R, args...; kwargs...)
lqr(Discrete, A, B, Q, R, args...; kwargs...)

Calculate the optimal gain matrix K for the state-feedback law u = -K*x that minimizes the cost function:

J = integral(x'Qx + u'Ru, 0, inf) for the continuous-time model dx = Ax + Bu. J = sum(x'Qx + u'Ru, 0, inf) for the discrete-time model x[k+1] = Ax[k] + Bu[k].

Solve the LQR problem for state-space system sys. Works for both discrete and continuous time systems.

The args...; kwargs... are sent to the Riccati solver, allowing specification of cross-covariance etc. See ?MatrixEquations.arec / ared for more help.

Examples

Continuous time

using LinearAlgebra # For identity matrix I
using Plots
A = [0 1; 0 0]
B = [0; 1]
C = [1 0]
sys = ss(A,B,C,0)
Q = I
R = I
L = lqr(sys,Q,R) # lqr(Continuous,A,B,Q,R) can also be used

u(x,t) = -L*x # Form control law,
t=0:0.1:5
x0 = [1,0]
y, t, x, uout = lsim(sys,u,t,x0=x0)
plot(t,x', lab=["Position" "Velocity"], xlabel="Time [s]")

Discrete time

using LinearAlgebra # For identity matrix I
using Plots
Ts = 0.1
A = [1 Ts; 0 1]
B = [0;1]
C = [1 0]
sys = ss(A, B, C, 0, Ts)
Q = I
R = I
L = lqr(Discrete, A,B,Q,R) # lqr(sys,Q,R) can also be used

u(x,t) = -L*x # Form control law,
t=0:Ts:5
x0 = [1,0]
y, t, x, uout = lsim(sys,u,t,x0=x0)
plot(t,x', lab=["Position"  "Velocity"], xlabel="Time [s]")

ControlSystems.place — Function

place(A, B, p, opt=:c)
place(sys::StateSpace, p, opt=:c)

Calculate the gain matrix K such that A - BK has eigenvalues p.

place(A, C, p, opt=:o)
place(sys::StateSpace, p, opt=:o)

Calculate the observer gain matrix L such that A - LC has eigenvalues p.

Uses Ackermann's formula. Currently handles only SISO systems.

ControlSystems.c2d — Function

sysd = c2d(sys::AbstractStateSpace{<:Continuous}, Ts, method=:zoh; w_prewarp=0)
Gd = c2d(G::TransferFunction{<:Continuous}, Ts, method=:zoh)

Convert the continuous-time system sys into a discrete-time system with sample time Ts, using the specified method (:zoh, :foh, :fwdeuler or :tustin). Note that the forward-Euler method generally requires the sample time to be very small relative to the time constants of the system.

method = :tustin performs a bilinear transform with prewarp frequency w_prewarp.

w_prewarp: Frequency (rad/s) for pre-warping when usingthe Tustin method, has no effect for other methods.

See also c2d_x0map

ControlSystems.c2d_poly2poly — Method

c2d_poly2poly(ro, Ts)

returns the polynomial coefficients in discrete time given polynomial coefficients in continuous time

ControlSystems.c2d_roots2poly — Method

c2d_roots2poly(ro, Ts)

returns the polynomial coefficients in discrete time given a vector of roots in continuous time

ControlSystems.c2d_x0map — Function

sysd, x0map = c2d_x0map(sys::AbstractStateSpace{<:Continuous}, Ts, method=:zoh; w_prewarp=0)

Returns the discretization sysd of the system sys and a matrix x0map that transforms the initial conditions to the discrete domain by x0_discrete = x0map*[x0; u0]

See c2d for further details.

ControlSystems.d2c — Function

d2c(sys::AbstractStateSpace{<:Discrete}, method::Symbol = :zoh; w_prewarp=0)

Convert discrete-time system to a continuous time system, assuming that the discrete-time system was discretized using method. Available methods are `:zoh, :fwdeuler´.

w_prewarp: Frequency for pre-warping when usingthe Tustin method, has no effect for other methods.

ControlSystems.dab — Method

X,Y = dab(A,B,C)

Solves the Diophantine-Aryabhatta-Bezout identity

$AX + BY = C$, where $A, B, C, X$ and $Y$ are polynomials and $deg Y = deg A - 1$.

See Computer-Controlled Systems: Theory and Design, Third Edition Karl Johan Åström, Björn Wittenmark

ControlSystems.rstc — Method

See ?rstd for the discrete case

ControlSystems.rstd — Method

R,S,T = rstd(BPLUS,BMINUS,A,BM1,AM,AO,AR,AS)
R,S,T = rstd(BPLUS,BMINUS,A,BM1,AM,AO,AR)
R,S,T = rstd(BPLUS,BMINUS,A,BM1,AM,AO)

Polynomial synthesis in discrete time.

Polynomial synthesis according to CCS ch 10 to design a controller $R(q) u(k) = T(q) r(k) - S(q) y(k)$

Inputs:

BPLUS : Part of open loop numerator

BMINUS : Part of open loop numerator

A : Open loop denominator

BM1 : Additional zeros

AM : Closed loop denominator

AO : Observer polynomial

AR : Pre-specified factor of R,

e.g integral part [1, -1]^k

AS : Pre-specified factor of S,

e.g notch filter [1, 0, w^2]

Outputs: R,S,T : Polynomials in controller

See function dab how the solution to the Diophantine- Aryabhatta-Bezout identity is chosen.

See Computer-Controlled Systems: Theory and Design, Third Edition Karl Johan Åström, Björn Wittenmark

ControlSystems.zpconv — Method

zpc(a,r,b,s)

form conv(a,r) + conv(b,s) where the lengths of the polynomials are equalized by zero-padding such that the addition can be carried out

ControlSystems.laglink — Method

laglink(a, M; [Ts])

Returns a phase retarding link, the rule of thumb a = 0.1ωc guarantees less than 6 degrees phase margin loss. The bode curve will go from M, bend down at a/M and level out at 1 for frequencies > a

ControlSystems.leadlink — Method

leadlink(b, N, K; [Ts])

Returns a phase advancing link, the top of the phase curve is located at ω = b√(N) where the link amplification is K√(N) The bode curve will go from K, bend up at b and level out at KN for frequencies > bN

The phase advance at ω = b√(N) can be plotted as a function of N with leadlinkcurve()

Values of N < 1 will give a phase retarding link.

See also leadlinkat laglink

ControlSystems.leadlinkat — Method

leadlinkat(ω, N, K; [Ts])

Returns a phase advancing link, the top of the phase curve is located at ω where the link amplification is K√(N) The bode curve will go from K, bend up at ω/√(N) and level out at KN for frequencies > ω√(N)

The phase advance at ω can be plotted as a function of N with leadlinkcurve()

Values of N < 1 will give a phase retarding link.

See also leadlink laglink

ControlSystems.leadlinkcurve — Function

leadlinkcurve(start=1)

Plot the phase advance as a function of N for a lead link (phase advance link) If an input argument s is given, the curve is plotted from s to 10, else from 1 to 10.

See also leadlink, leadlinkat

ControlSystems.loopshapingPI — Method

C, kp, ki, fig = loopshapingPI(P, ωp; ϕl, rl, phasemargin, form=:standard, doplot=false)

Selects the parameters of a PI-controller (on parallel form) such that the Nyquist curve of P at the frequency ωp is moved to rl exp(i ϕl)

The parameters can be returned as one of several common representations chosen by form, the options are

:standard - Kp*(1 + 1/(Ti*s) + Td*s)

:series - Kc*(1 + 1/(τi*s))*(τd*s + 1)

:parallel - Kp + Ki/s + Kd*s

If phasemargin is supplied (in degrees), ϕl is selected such that the curve is moved to an angle of phasemargin - 180 degrees

If no rl is given, the magnitude of the curve at ωp is kept the same and only the phase is affected, the same goes for ϕl if no phasemargin is given.

Set doplot = true to plot the gangoffourplot and nyquistplot of the system.

See also pidplots, stabregionPID and placePI.

ControlSystems.pid — Function

C = pid(param_p, param_i, [param_d]; form=:standard, state_space=false, [Tf], [Ts])

Calculates and returns a PID controller.

The form can be chosen as one of the following

:standard - Kp*(1 + 1/(Ti*s) + Td*s)

:series - Kc*(1 + 1/(τi*s))*(τd*s + 1)

:parallel - Kp + Ki/s + Kd*s

If state_space is set to true, either kd has to be zero or a positive Tf has to be provided for creating a filter on the input to allow for a state space realization. The filter used is 1 / (1 + s*Tf + (s*Tf)^2/2), where Tf can typically be chosen as Ti/N for a PI controller and Td/N for a PID controller, and N is commonly in the range 2 to 20. The state space will be returned on controllable canonical form.

For a discrete controller a positive Ts can be supplied.

Examples

C1 = pid(3.3, 1, 2)                             # Kd≠0 works without filter in tf form
C2 = pid(3.3, 1, 2; Tf=0.3, state_space=true)   # In statespace a filter is needed
C3 = pid(2., 3, 0; Ts=0.4, state_space=true)    # Discrete

The functions pid_tf and pid_ss are also exported. They take the same parameters and is what is actually called in pid based on the state_space parameter.

ControlSystems.pidplots — Method

pidplots(P, args...; params_p, params_i, params_d=0, form=:standard, ω=0, grid=false, kwargs...)

Plots interesting figures related to closing the loop around process P with a PID controller supplied in params on one of the following forms:

:standard - Kp*(1 + 1/(Ti*s) + Td*s)

:series - Kc*(1 + 1/(τi*s))*(τd*s + 1)

:parallel - Kp + Ki/s + Kd*s

The sent in values can be arrays to evaluate multiple different controllers, and if grid=true it will be a grid search over all possible combinations of the values.

Available plots are :gof for Gang of four, :nyquist, :controller for a bode plot of the controller TF and :pz for pole-zero maps and should be supplied as additional arguments to the function.

One can also supply a frequency vector ω to be used in Bode and Nyquist plots.

See also loopshapingPI, stabregionPID

ControlSystems.placePI — Method

C, p, i = placePI(P, ω₀, ζ; form=:standard)

Selects the parameters of a PI-controller such that the poles of closed loop between P and C are placed to match the poles of s^2 + 2ζω₀s + ω₀^2.

The parameters can be returned as one of several common representations chose by form, the options are

:standard - Kp*(1 + 1/(Ti*s))

:series - Kc*(1 + 1/(τi*s)) (equivalent to above for PI controllers)

:parallel - Kp + Ki/s

C is the returned transfer function of the controller and params is a named tuple containing the parameters. The parameters can be accessed as params.Kp or params["Kp"] from the named tuple, or they can be unpacked using Kp, Ti, Td = values(params).

See also loopshapingPI

ControlSystems.rlocusplot — Function

rlocusplot(P::LTISystem, K)

Computes and plots the root locus of the SISO LTISystem P with a negative feedback loop and feedback gains K, if K is not provided, range(1e-6,stop=500,length=10000) is used. If OrdinaryDiffEq.jl is installed and loaded by the user (using OrdinaryDiffEq), rlocusplot will use an adaptive step-size algorithm to select values of K. A scalar Kmax can then be given as second argument.

ControlSystems.stabregionPID — Function

kp, ki, fig = stabregionPID(P, [ω]; kd=0, doplot=false, form=:standard)

Segments of the curve generated by this program is the boundary of the stability region for a process with transfer function P(s) The provided derivative gain is expected on parallel form, i.e., the form kp + ki/s + kd s, but the result can be transformed to any form given by the form keyword. The curve is found by analyzing P(s)*C(s) = -1 ⟹ |PC| = |P| |C| = 1 arg(P) + arg(C) = -π If P is a function (e.g. s -> exp(-sqrt(s)) ), the stability of feedback loops using PI-controllers can be analyzed for processes with models with arbitrary analytic functions See also stabregionPID, loopshapingPI, pidplots

ControlSystems.sminreal — Method

sminreal(sys)

Compute the structurally minimal realization of the state-space system sys. A structurally minimal realization is one where only states that can be determined to be uncontrollable and unobservable based on the location of 0s in sys are removed.

Systems with numerical noise in the coefficients, e.g., noise on the order of eps require truncation to zero to be affected by structural simplification, e.g.,

trunc_zero!(A) = A[abs.(A) .< 10eps(maximum(abs, A))] .= 0
trunc_zero!(sys.A); trunc_zero!(sys.B); trunc_zero!(sys.C)
sminreal(sys)

See also minreal

ControlSystems.add_input — Function

add_input(sys::AbstractStateSpace, B2::AbstractArray, D2 = 0)

Add inputs to sys by forming

\[x' = Ax + [B B2]u y = Cx + [D D2]u\]

If B2 is an integer it will be interpreted as an index and an input matrix containing a single 1 at the specified index will be used.

Example: The following example forms an innovation model that takes innovations as inputs

G   = ssrand(2,2,3, Ts=1)
K   = kalman(G, I(G.nx), I(G.ny))
sys = add_input(G, K)

ControlSystems.add_output — Function

add_output(sys::AbstractStateSpace, C2::AbstractArray, D2 = 0)

Add outputs to sys by forming

\[x' = Ax + Bu y = [C; C2]x + [D; D2]u\]

If C2 is an integer it will be interpreted as an index and an output matrix containing a single 1 at the specified index will be used.

ControlSystems.append — Method

append(systems::StateSpace...), append(systems::TransferFunction...)

Append systems in block diagonal form

ControlSystems.array2mimo — Method

array2mimo(M::AbstractArray{<:LTISystem})

Take an array of LTISystems and create a single MIMO system.

ControlSystems.feedback — Method

feedback(sys1::AbstractStateSpace, sys2::AbstractStateSpace;
         U1=:, Y1=:, U2=:, Y2=:, W1=:, Z1=:, W2=Int[], Z2=Int[],
         Wperm=:, Zperm=:, pos_feedback::Bool=false)

Basic use feedback(sys1, sys2) forms the feedback interconnection

           ┌──────────────┐
◄──────────┤     sys1     │◄──── Σ ◄──────
    │      │              │      │
    │      └──────────────┘      -1
    │                            |
    │      ┌──────────────┐      │
    └─────►│     sys2     ├──────┘
           │              │
           └──────────────┘

Advanced use feedback also supports more flexible use according to the figure below

              ┌──────────────┐
      z1◄─────┤     sys1     │◄──────w1
 ┌─── y1◄─────┤              │◄──────u1 ◄─┐
 │            └──────────────┘            │
 │                                        α
 │            ┌──────────────┐            │
 └──► u2─────►│     sys2     ├───────►y2──┘
      w2─────►│              ├───────►z2
              └──────────────┘

U1, W1 specifies the indices of the input signals of sys1 corresponding to u1 and w1 Y1, Z1 specifies the indices of the output signals of sys1 corresponding to y1 and z1 U2, W2, Y2, Z2 specifies the corresponding signals of sys2

Specify Wperm and Zperm to reorder the inputs (corresponding to [w1; w2]) and outputs (corresponding to [z1; z2]) in the resulting statespace model.

Negative feedback (α = -1) is the default. Specify pos_feedback=true for positive feedback (α = 1).

See also lft, starprod, sensitivity, input_sensitivity, output_sensitivity, comp_sensitivity, input_comp_sensitivity, output_comp_sensitivity, G_PS, G_CS.

See Zhou, Doyle, Glover (1996) for similar (somewhat less symmetric) formulas.

ControlSystems.feedback — Method

feedback(sys)
feedback(sys1, sys2)

For a general LTI-system, feedback forms the negative feedback interconnection

>-+ sys1 +-->
  |      |
 (-)sys2 +

If no second system is given, negative identity feedback is assumed

ControlSystems.feedback2dof — Method

feedback2dof(P,R,S,T)
feedback2dof(B,A,R,S,T)

Return BT/(AR+ST) where B and A are the numerator and denomenator polynomials of P respectively

Return BT/(AR+ST)

ControlSystems.feedback2dof — Method

feedback2dof(P::TransferFunction, C::TransferFunction, F::TransferFunction)

Return the transfer function P(F+C)/(1+PC) which is the closed-loop system with process P, controller C and feedforward filter F from reference to control signal (by-passing C).

         +-------+
         |       |
   +----->   F   +----+
   |     |       |    |
   |     +-------+    |
   |     +-------+    |    +-------+
r  |  -  |       |    |    |       |    y
+--+----->   C   +----+---->   P   +---+-->
      |  |       |         |       |   |
      |  +-------+         +-------+   |
      |                                |
      +--------------------------------+

ControlSystems.lft — Function

lft(G, Δ, type=:l)

Lower and upper linear fractional transformation between systems G and Δ.

Specify :l lor lower LFT, and :u for upper LFT.

G must have more inputs and outputs than Δ has outputs and inputs.

For details, see Chapter 9.1 in Zhou, K. and JC Doyle. Essentials of robust control, Prentice hall (NJ), 1998

ControlSystems.parallel — Method

parallel(sys1::LTISystem, sys2::LTISystem)

Connect systems in parallel, equivalent to sys2+sys1

ControlSystems.series — Method

series(sys1::LTISystem, sys2::LTISystem)

Connect systems in series, equivalent to sys2*sys1

ControlSystems.starprod — Method

starprod(sys1, sys2, dimu, dimy)

Compute the Redheffer star product.

length(U1) = length(Y2) = dimu and length(Y1) = length(U2) = dimy

For details, see Chapter 9.3 in Zhou, K. and JC Doyle. Essentials of robust control, Prentice hall (NJ), 1998

ControlSystems.G_CS — Method

G_CS(P, C)

The closed-loop transfer function from (-) measurement noise or (+) reference to control signal. Technically, the transfer function is given by (1 + CP)⁻¹C so SC would be a better, but nonstandard name.

         ▲
         │e₁
         │  ┌─────┐
d₁────+──┴──►  P  ├─────┬──►e₄
      │     └─────┘     │
      │                 │
      │     ┌─────┐    -│
 e₂◄──┴─────┤  C  ◄──┬──+───d₂
            └─────┘  │
                     │e₃
                     ▼

input_sensitivity is the transfer function from d₁ to e₁, (I + CP)⁻¹

output_sensitivity is the transfer function from d₂ to e₃, (I + PC)⁻¹

input_comp_sensitivity is the transfer function from d₁ to e₂, (I + CP)⁻¹CP

output_comp_sensitivity is the transfer function from d₂ to e₄, (I + PC)⁻¹PC

G_PS is the transfer function from d₁ to e₄, (1 + PC)⁻¹P

G_CS is the transfer function from d₂ to e₂, (1 + CP)⁻¹C

ControlSystems.G_PS — Method

G_PS(P, C)

The closed-loop transfer function from load disturbance to plant output. Technically, the transfer function is given by (1 + PC)⁻¹P so SP would be a better, but nonstandard name.

         ▲
         │e₁
         │  ┌─────┐
d₁────+──┴──►  P  ├─────┬──►e₄
      │     └─────┘     │
      │                 │
      │     ┌─────┐    -│
 e₂◄──┴─────┤  C  ◄──┬──+───d₂
            └─────┘  │
                     │e₃
                     ▼

input_sensitivity is the transfer function from d₁ to e₁, (I + CP)⁻¹

output_sensitivity is the transfer function from d₂ to e₃, (I + PC)⁻¹

input_comp_sensitivity is the transfer function from d₁ to e₂, (I + CP)⁻¹CP

output_comp_sensitivity is the transfer function from d₂ to e₄, (I + PC)⁻¹PC

G_PS is the transfer function from d₁ to e₄, (1 + PC)⁻¹P

G_CS is the transfer function from d₂ to e₂, (1 + CP)⁻¹C

ControlSystems.comp_sensitivity — Method

See output_comp_sensitivity

         ▲
         │e₁
         │  ┌─────┐
d₁────+──┴──►  P  ├─────┬──►e₄
      │     └─────┘     │
      │                 │
      │     ┌─────┐    -│
 e₂◄──┴─────┤  C  ◄──┬──+───d₂
            └─────┘  │
                     │e₃
                     ▼

input_sensitivity is the transfer function from d₁ to e₁, (I + CP)⁻¹

output_sensitivity is the transfer function from d₂ to e₃, (I + PC)⁻¹

input_comp_sensitivity is the transfer function from d₁ to e₂, (I + CP)⁻¹CP

output_comp_sensitivity is the transfer function from d₂ to e₄, (I + PC)⁻¹PC

G_PS is the transfer function from d₁ to e₄, (1 + PC)⁻¹P

G_CS is the transfer function from d₂ to e₂, (1 + CP)⁻¹C

ControlSystems.input_comp_sensitivity — Method

input_comp_sensitivity(P,C)

Transfer function from load disturbance to control signal.

"Input" signifies that the transfer function is from the input of the plant.

"Complimentary" signifies that the transfer function is to an output (in this case controller output)

         ▲
         │e₁
         │  ┌─────┐
d₁────+──┴──►  P  ├─────┬──►e₄
      │     └─────┘     │
      │                 │
      │     ┌─────┐    -│
 e₂◄──┴─────┤  C  ◄──┬──+───d₂
            └─────┘  │
                     │e₃
                     ▼

input_sensitivity is the transfer function from d₁ to e₁, (I + CP)⁻¹

output_sensitivity is the transfer function from d₂ to e₃, (I + PC)⁻¹

input_comp_sensitivity is the transfer function from d₁ to e₂, (I + CP)⁻¹CP

output_comp_sensitivity is the transfer function from d₂ to e₄, (I + PC)⁻¹PC

G_PS is the transfer function from d₁ to e₄, (1 + PC)⁻¹P

G_CS is the transfer function from d₂ to e₂, (1 + CP)⁻¹C

ControlSystems.input_sensitivity — Method

input_sensitivity(P, C)

Transfer function from load disturbance to total plant input.

"Input" signifies that the transfer function is from the input of the plant.

         ▲
         │e₁
         │  ┌─────┐
d₁────+──┴──►  P  ├─────┬──►e₄
      │     └─────┘     │
      │                 │
      │     ┌─────┐    -│
 e₂◄──┴─────┤  C  ◄──┬──+───d₂
            └─────┘  │
                     │e₃
                     ▼

input_sensitivity is the transfer function from d₁ to e₁, (I + CP)⁻¹

output_sensitivity is the transfer function from d₂ to e₃, (I + PC)⁻¹

input_comp_sensitivity is the transfer function from d₁ to e₂, (I + CP)⁻¹CP

output_comp_sensitivity is the transfer function from d₂ to e₄, (I + PC)⁻¹PC

G_PS is the transfer function from d₁ to e₄, (1 + PC)⁻¹P

G_CS is the transfer function from d₂ to e₂, (1 + CP)⁻¹C

ControlSystems.output_comp_sensitivity — Method

output_comp_sensitivity(P,C)

Transfer function from measurement noise / reference to plant output.

"output" signifies that the transfer function is from the output of the plant.

"Complimentary" signifies that the transfer function is to an output (in this case plant output)

         ▲
         │e₁
         │  ┌─────┐
d₁────+──┴──►  P  ├─────┬──►e₄
      │     └─────┘     │
      │                 │
      │     ┌─────┐    -│
 e₂◄──┴─────┤  C  ◄──┬──+───d₂
            └─────┘  │
                     │e₃
                     ▼

input_sensitivity is the transfer function from d₁ to e₁, (I + CP)⁻¹

output_sensitivity is the transfer function from d₂ to e₃, (I + PC)⁻¹

input_comp_sensitivity is the transfer function from d₁ to e₂, (I + CP)⁻¹CP

output_comp_sensitivity is the transfer function from d₂ to e₄, (I + PC)⁻¹PC

G_PS is the transfer function from d₁ to e₄, (1 + PC)⁻¹P

G_CS is the transfer function from d₂ to e₂, (1 + CP)⁻¹C

ControlSystems.output_sensitivity — Method

output_sensitivity(P, C)

Transfer function from measurement noise / reference to control error.

"output" signifies that the transfer function is from the output of the plant.

         ▲
         │e₁
         │  ┌─────┐
d₁────+──┴──►  P  ├─────┬──►e₄
      │     └─────┘     │
      │                 │
      │     ┌─────┐    -│
 e₂◄──┴─────┤  C  ◄──┬──+───d₂
            └─────┘  │
                     │e₃
                     ▼

input_sensitivity is the transfer function from d₁ to e₁, (I + CP)⁻¹

output_sensitivity is the transfer function from d₂ to e₃, (I + PC)⁻¹

input_comp_sensitivity is the transfer function from d₁ to e₂, (I + CP)⁻¹CP

output_comp_sensitivity is the transfer function from d₂ to e₄, (I + PC)⁻¹PC

G_PS is the transfer function from d₁ to e₄, (1 + PC)⁻¹P

G_CS is the transfer function from d₂ to e₂, (1 + CP)⁻¹C

ControlSystems.sensitivity — Method

See output_sensitivity

         ▲
         │e₁
         │  ┌─────┐
d₁────+──┴──►  P  ├─────┬──►e₄
      │     └─────┘     │
      │                 │
      │     ┌─────┐    -│
 e₂◄──┴─────┤  C  ◄──┬──+───d₂
            └─────┘  │
                     │e₃
                     ▼

input_sensitivity is the transfer function from d₁ to e₁, (I + CP)⁻¹

output_sensitivity is the transfer function from d₂ to e₃, (I + PC)⁻¹

input_comp_sensitivity is the transfer function from d₁ to e₂, (I + CP)⁻¹CP

output_comp_sensitivity is the transfer function from d₂ to e₄, (I + PC)⁻¹PC

G_PS is the transfer function from d₁ to e₄, (1 + PC)⁻¹P

G_CS is the transfer function from d₂ to e₂, (1 + CP)⁻¹C

ControlSystems.bodev — Method

bodev(sys::LTISystem, w::AbstractVector; $(Expr(:kw, :unwrap, true)))

For use with SISO systems where it acts the same as bode but with the extra dimensions removed in the returned values.

ControlSystems.bodev — Method

bodev(sys::LTISystem; )

For use with SISO systems where it acts the same as bode but with the extra dimensions removed in the returned values.

ControlSystems.freqrespv — Method

freqrespv(G::AbstractMatrix, w_vec::AbstractVector{<:Real}; )

For use with SISO systems where it acts the same as freqresp but with the extra dimensions removed in the returned values.

ControlSystems.freqrespv — Method

freqrespv(G::Number, w_vec::AbstractVector{<:Real}; )

For use with SISO systems where it acts the same as freqresp but with the extra dimensions removed in the returned values.

ControlSystems.freqrespv — Method

freqrespv(sys::LTISystem, w_vec::AbstractVector{W}; )

For use with SISO systems where it acts the same as freqresp but with the extra dimensions removed in the returned values.

ControlSystems.nyquistv — Method

nyquistv(sys::LTISystem, w::AbstractVector; )

For use with SISO systems where it acts the same as nyquist but with the extra dimensions removed in the returned values.

ControlSystems.nyquistv — Method

nyquistv(sys::LTISystem; )

For use with SISO systems where it acts the same as nyquist but with the extra dimensions removed in the returned values.

ControlSystems.sigmav — Method

sigmav(sys::LTISystem, w::AbstractVector; )

For use with SISO systems where it acts the same as sigma but with the extra dimensions removed in the returned values.

ControlSystems.sigmav — Method

sigmav(sys::LTISystem; )

For use with SISO systems where it acts the same as sigma but with the extra dimensions removed in the returned values.